% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_net_clogit.R
\name{bag_fit_net_clogit}
\alias{bag_fit_net_clogit}
\title{Fit a bag of conditional logistic regression/SSF/iSSF models with penalized regression in a train-validate-test setup}
\usage{
bag_fit_net_clogit(
  f,
  data,
  samples,
  kernel_vars = c("step_length", "ta"),
  metric = c(conditionalBoyce, somersD, AUC)[[1]],
  out_dir_file = NULL,
  na.action = "na.pass",
  parallel = c(FALSE, "foreach", "mclapply")[1],
  mc.cores = 2L,
  ...
)
}
\arguments{
\item{mc.cores}{Only relevant if \code{parallel == "mclapply"}. If \code{parallel == "foreach"}, cores must
be assigned before running \code{fit_multi_net_logit()} using \code{\link[parallel:makeCluster]{parallel::makeCluster()}} and
\code{\link[doParallel:registerDoParallel]{doParallel::registerDoParallel()}}.}

\item{...}{Options for net_logit and glmnet}
}
\description{
Fit a bag of conditional logistic regression/SSF/iSSF models with penalized regression in a train-validate-test setup
}
