% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_net_clogit.R
\name{bag_load_net_clogit}
\alias{bag_load_net_clogit}
\title{Load a series of files output of fit_net_clogit models and put them on a bag}
\usage{
bag_load_net_clogit(
  f,
  data,
  samples,
  load_models_path = ".",
  load_models_pattern = NULL,
  metric = c(conditionalBoyce, conditionalSomersD, conditionalAUC)[[1]],
  standardize = c("internal", "external", FALSE)[1],
  method = c("Lasso", "Ridge", "AdaptiveLasso", "DecayAdaptiveLasso", "ElasticNet")[1]
)
}
\arguments{
\item{load_models_path}{\verb{[string="."]} \cr Path to the folder where the files
are saved.}

\item{load_models_pattern}{\verb{[string="."]} \cr Pattern common to the file names,
to be used in \code{grep()} to find the files within the folder
\code{load_models_path}. It should be the pattern present in the argument
\code{out_dir_file} in \code{\link[=fit_net_clogit]{fit_net_clogit()}} or \code{\link[=bag_fit_net_clogit]{bag_fit_net_clogit()}}
functions.}
}
\description{
Load a series of files output of fit_net_clogit models and put them on a bag
}
