% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_plot_response.R
\name{bag_plot_response}
\alias{bag_plot_response}
\title{Plot responses from a bag of models}
\usage{
bag_plot_response(
  x,
  dfvar,
  data,
  type = c("linear", "exponential")[2],
  zoi_shape = c("exp_decay", "gaussian_decay", "linear_decay", "threshold_decay")[1],
  ci = TRUE,
  wQ_probs = c(0.025, 0.5, 0.975),
  baseline = NULL,
  zoi = FALSE,
  zoi_vals = c(100, 250, 500, 1000, 2500, 5000, 10000),
  ggplot = T,
  plot_mean = TRUE,
  plot_median = TRUE,
  normalize = c(FALSE, "mean", "median", "ci")[1],
  logx = F,
  ylim = NULL
)
}
\arguments{
\item{x}{\verb{[list]} \cr A bag of models, resulting from a call to \code{\link[=bag_models]{bag_models()}}.}

\item{ci}{Should variation or confidence intervals be plotted?}

\item{wQ_probs}{a three element with lower, mid, and higher weighted quantiles to be computed}
}
\description{
Plot responses from a bag of models
}
