% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cabins.gpkg}
\alias{cabins.gpkg}
\title{Cabins vector data for the reindeer area}
\format{
A geopackage file. Projected CRS: ETRS89 / UTM zone 33N. The vector
file presents the following columns:
\itemize{
\item{gid:} {Line number, corresponding to the original dataset}
\item{buildtype:} {Type of building (code) in the original dataset}
\item{city:} {Code of the municipality where the cabin is located}
\item{value:} {Value 1, to be used for rasterization purposes}
}
}
\source{
\url{https://register.geonorge.no/det-offentlige-kartgrunnlaget/n50-kartdata/ea192681-d039-42ec-b1bc-f3ce04c189ac}
}
\description{
Dataset containing the location of tourist private cabins is Southern
Norway, within the reindeer management area of Setesdal Austhei.
It corresponds to some specific building types (object_type = "Bygning",
byggtyp_nbr = c("161", "162", "163")) from the public N50 dataset.
}
\examples{
(s <- system.file("vector/cabins.gpkg", package = "oneimpact"))
sf::st_read(s)
# or
terra::vect(s)

}
