% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cabins_count.tif}
\alias{cabins_count.tif}
\title{Cabin count raster data}
\format{
A Geotiff file. Projected CRS: ETRS89 / UTM zone 33N.
}
\source{
\url{https://register.geonorge.no/det-offentlige-kartgrunnlaget/n50-kartdata/ea192681-d039-42ec-b1bc-f3ce04c189ac}
}
\description{
Raster data indicating the number of tourist private cabins per pixel in Norway.
It corresponds to some specific building types (object_type = "Bygning",
byggtyp_nbr = c("161", "162", "163")) form the public N50 dataset.
The original data consisted of point vector data and were rasterized with 100m
resolution by counting the number of cabins in each pixel. The raster
was cut for the study area presented in the \link{oneimpact} package.
}
\examples{
(f <- system.file("raster/cabins_count.tif", package = "oneimpact"))
terra::rast(f)

}
