% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cabins_vect.gpkg}
\alias{cabins_vect.gpkg}
\title{Cabins vector data}
\format{
A geopackage file. Projected CRS: ETRS89 / UTM zone 33N. The vector
file presents the following columns:
\itemize{
\item{cat:} {Line number, corresponding to the original dataset}
\item{byggtyp_nbr:} {Type of building (code) in the original dataset}
\item{kommune:} {Code of the municipality where the cabin is located}
\item{value:} {Value 1, to be used for rasterization purposes}
}
}
\source{
\url{to_add_here}
}
\description{
Dataset containing the location of tourist private cabins is Southern
Norway, within the study area for the \link{oneimpact} package.
It corresponds to
some specific building types (byggtyp_nbr = c("161", "162", "163",
"171", "172")) form the public N50 dataset (describe further here).
}
\examples{
(s <- system.file("vector/cabins_vect.gpkg", package = "oneimpact"))
sf::st_read(s)
# or
terra::vect(s)

}
