% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_boyce.R
\name{conditionalBoyce}
\alias{conditionalBoyce}
\title{Computes de conditional boyce index for model evaluation}
\usage{
conditionalBoyce(
  x,
  method = c("pearson", "kendall", "spearman")[1],
  plotit = FALSE,
  errors = TRUE,
  warnings = TRUE
)
}
\description{
x: dataframe with x: predicted value, y:use vs. available, strat: stratum
}
