% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_predict.R
\name{create_linear_feature_zoi}
\alias{create_linear_feature_zoi}
\title{The function \code{create_linear_feature_zoi} created ZOI of one single linear feature so it is used to
correctly create response plots for linear infrastructure, considering the potential
responses at multiple radii.}
\usage{
create_linear_feature_zoi(
  radii = c(100, 250, 500, 1000, 2500, 5000, 10000),
  type = c("circle", "Gauss", "rectangle", "exp_decay", "bartlett", "threshold",
    "mfilter")[1],
  radius_max = 10000,
  res = 100,
  value = 1
)
}
\description{
The function \code{create_linear_feature_zoi} created ZOI of one single linear feature so it is used to
correctly create response plots for linear infrastructure, considering the potential
responses at multiple radii.
}
