% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_filter.R
\name{ed}
\alias{ed}
\title{Create filters for raster neighborhood analyses}
\usage{
ed(
  half_life = 1000,
  method = c("exp_decay"),
  max_dist = 5000,
  res = 100,
  save_txt = FALSE
)
}
\value{
A matrix with the weight values.
}
\description{
This function creates matrices of weights following different
functions to be used in neighborhood analyses for rasters. It is possible
to export these matrices as text files, for use with external softwares
such as the r.mfilter module within GRASS GIS.
}
\examples{

}
