% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_blocks.R
\name{explore_blocks}
\alias{explore_blocks}
\alias{explore_blocks_pre}
\title{Explore hierarchical blocks before and after spatial stratification}
\usage{
explore_blocks_pre(data, colH0, animal_id = NULL, col_case = "case")

explore_blocks(blocks)
}
\arguments{
\item{data}{\verb{[data.frame]} \cr Data.}

\item{colH0}{\verb{[character]} \cr Name of the column to be used as the H0 level.}
}
\description{
Functions to explore the number of cases and observations for the different
sampling units possibly used as the base H0 hierarchical level, such as population ID,
study area, animal ID, or year.
}
