% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_blocks.R
\name{explore_blocks_pre}
\alias{explore_blocks_pre}
\title{Explore potential hierarchical blocks before sampling or spatial stratification}
\usage{
explore_blocks_pre(data, colH0, animal_id = NULL, col_case = "case")
}
\arguments{
\item{data}{\verb{[data.frame,tibble]} \cr Complete data set to be analyzed.}

\item{colH0}{\verb{[character]} \cr Name of the column in \code{data} to be used as the H0
hierarchical level, intended for model validation.}

\item{animal_id}{\verb{[character]} \cr Name of the column in \code{data} representing
animal ID. If \code{NULL} (default), summaries are not created for individuals.}

\item{col_case}{\verb{[string(1)="case"]} \cr Name of the column in \code{data} representing
the case or used/available points. Default is \code{"case"}.}
}
\value{
A \code{data.frame} with one row per unique combination of \code{colH0} (and
\code{animal_id} if provided), with columns:
\itemize{
\item \code{n}: total number of observations in the group.
\item \code{n_presences}: number of used/presence points (\code{col_case == 1}) in the group.
}
}
\description{
Function to explore the number of cases and observations for the different
sampling units possibly used as the base H0 hierarchical level, such as population ID,
study area, animal ID, or year, before spatial stratification or creating samples
for the bootstrapped approach.
The function can help understand how imbalanced is the data across H0 levels used
for validation.
}
\examples{
# read data
data("reindeer_ssf")

# explore blocks - animal ID as block H0
explore_blocks_pre(reindeer_ssf, "original_animal_id", col_case = "case_")

# explore blocks - year as block H0
library(lubridate)
reindeer_ssf |>
  dplyr::mutate(year = lubridate::year(t1_)) |>
  explore_blocks_pre("year", col_case = "case_")

# year as block H0 + animal ID
reindeer_ssf |>
  dplyr::mutate(year = lubridate::year(t1_)) |>
  explore_blocks_pre("year", animal_id = "original_animal_id", col_case = "case_")
}
