% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_case_strata.R
\name{extract_case_strata}
\alias{extract_case_strata}
\title{Separates the reponse variable, the strata variable, and other variables in a statistical formula}
\usage{
extract_case_strata(f, other_vars = F)
}
\arguments{
\item{f}{Formula}

\item{other_vars}{logical. If true, all the (other) explanatory variables are also
appended to the output, separated from the responde and strata variables.}
}
\value{
A \code{data.frame} with the response variable (\code{case}), the stratum variable
(\code{strata}), and possibly all explanatory variables, if \code{other_vars = TRUE}.
}
\description{
Separates the reponse variable, the strata variable, and other variables in a statistical formula
}
\examples{
f <- formula(use ~ exp1 + exp2 + exp1:exp2 + strata(id))
extract_case_strata(f)
extract_case_strata(f, other_vars = TRUE)

}
