% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_na_strata.R
\name{filter_na_strata}
\alias{filter_na_strata}
\title{Remove missing values and ensure case-control per stratum}
\usage{
filter_na_strata(f, data)
}
\arguments{
\item{f}{formula}

\item{data}{dataset}
}
\value{
Cleaned \code{data.frame}, removing from the input \code{data} the rows with NA
in any of the columns and all the strata for which there are only presences or
only absences.
}
\description{
Remove strata for which there are only 1s (used points, presence) or only 0s
(available points, background or absence).
}
\examples{
library(survival)
# create test dataset
test1 <- data.frame(case = c(0,0,1,0,1,1,0,1,NA,1),
                    x = c(0,2,1,2,1,0,1,0,0,1),
                    step_id = c(0,0,0,0,1,1,1,2,2,2))
# Remove NAs
f <- case ~ x + strata(step_id)
filter_na_strata(f, test1)

# Fit a stratified model
if(FALSE) {
  # with no NAs; necessary if using glmnet
  coxph(Surv(rep(1, length(case)), case) ~ x + strata(step_id), filter_na_strata(f, test1))
  # This differs from that
  coxph(Surv(rep(1, length(case)), case) ~ x + strata(step_id), )
}
}
\references{
\code{\link[amt:remove_incomplete_strata]{amt::remove_incomplete_strata()}}
}
