% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_find_layers.R
\name{grass_find_layer}
\alias{grass_find_layer}
\title{Find layers within GRASS GIS with multiple patterns}
\usage{
grass_find_layer(
  list_patterns,
  layers_grass = NULL,
  type = "raster",
  pattern = "*",
  mapset = "PERMANENT"
)
}
\arguments{
\item{list_patterns}{\verb{[list,character]} \cr List of strings, each corresponding to a pattern
to be searched in the names of layers within a GRASS GIS mapset. The patterns are filtered
one after the other, in the order they are listed.}

\item{layers_grass}{\verb{[vector,character]} \cr Vector of strings with the names of maps being
assessed, within a GRASS GIS mapset, such as the ones created through the \code{g.list} module.
If \code{NULL} (default), the list of maps within the GRASS GIS mapset is assessed within the function.}

\item{type}{\verb{[character(1)="raster"]} \cr Type of layer to be listed within the GRASS GIS mapset
(e.g. "raster", "vector"), if \code{layers_grass} is \code{NULL}.}

\item{pattern}{\verb{[character]} \cr Regular expression used to list maps within the GRASS GIS mapset,
if \code{layers_grass} is \code{NULL}. Default is "*", so all maps of a given \code{type} are listed.}

\item{mapset}{\verb{[character(1)="PERMANENT"]} \cr Name of the mapset from which maps are listed,
if \code{layers_grass} is \code{NULL}. Default is "PERMANENT".}
}
\value{
One or more strings with the names of the maps within the GRASS GIS mapset.
}
\description{
Find layers within GRASS GIS with multiple patterns
}
