% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hardanger_cabins_public.gpkg}
\alias{hardanger_cabins_public.gpkg}
\title{Public cabins vector data for the Hardangervidda wild reindeer area in Norway}
\format{
A geopackage file. Projected CRS: ETRS89 / UTM zone 33N. The vector
file presents the following columns:
\itemize{
\item{gid:} {Line number, corresponding to the original dataset}
\item{buildtype:} {Type of building (code) in the original dataset}
\item{city:} {Code of the municipality where the cabin is located}
\item{value:} {Value 1, to be used for rasterization purposes}
}
}
\source{
\url{https://register.geonorge.no/det-offentlige-kartgrunnlaget/n50-kartdata/ea192681-d039-42ec-b1bc-f3ce04c189ac}
}
\description{
Dataset containing the location of large, public DNT cabins in the
surroundings of the Hardangervidda wild reindeer area.
Retrieved from the public N50 dataset.
}
\examples{
(f <- system.file("vector/hardanger_cabins_public.gpkg", package = "oneimpact"))
v <- sf::st_read(f)
plot(v[1])

}
