% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hardanger_polygon.gpkg}
\alias{hardanger_polygon.gpkg}
\title{Polygon delimiting the study area around the Hardangervidda wild reindeer area}
\format{
A geopackage file. Projected CRS: ETRS89 / UTM zone 33N. The vector
file presents the following columns:
\itemize{
\item{reindeer_areas_id:} {ID of the area}
\item{name_area:} {Name of the wild reindeer area}
}
}
\source{
Niebuhr, B. B., Van Moorter, B., Stien, A., Tveraa, T., Strand, O., Langeland, K.,
Sandström, P., Alam, M., Skarin, A., & Panzacchi, M. (2023). Estimating the cumulative impact
and zone of influence of anthropogenic features on biodiversity.
Methods in Ecology and Evolution. https://doi.org/10.1111/2041-210X.14133
}
\description{
Dataset containing the polygon of the Hardangervidda area. It is different from
the official wild reindeer area and is delimited by the main roads and barriers in the surroundings
of the Hardangervidda plateau in Southern Norway,
}
\examples{
(f <- system.file("vector/hardanger_polygon.gpkg", package = "oneimpact"))
v <- terra::vect(f)
plot(v)

}
