% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hardanger_rast_predictors_500m.tif}
\alias{hardanger_rast_predictors_500m.tif}
\title{Predictor environmental variables for the Hardangervidda wild reindeer area in Norway}
\format{
A Geotiff file. Projected CRS: ETRS89 / UTM zone 33N.
\itemize{
\item{norway_pca_klima_axis1-4}{Components 1 to 4 from a principal component analysis
representing bio-geo-climatic variation in Norway, from Bakkestuen et al. 2008.
PCAs 1 to 4 represent, respectively, continentality, altitude, terrain ruggedness, and solar radiation.
More information in Niebuhr et al. 2023. PCAs 1 and 2 also present quadratic layers.}
\item{NORUTreclass} {Land use and land cover classes from NORUT, reclassified as in Niebuhr et al. 2023.}
\item{cabins_private_nearest_exp_decayXXX} {Zone of influence of the nearest private cabin at each location,
with exponential decay shape, and radii defined by XXX (from 100 to 10000 m).}
\item{cabins_private_cumulative_exp_decayXXX} {Cumulative zone of influence of private cabins at each location,
with exponential decay shape, and radii defined by XXX (from 100 to 10000 m).}
\item{cabins_public_nearest_exp_decayXXX} {Zone of influence of the nearest public resort and mountain hotels
at each location, with exponential decay shape, and radii defined by XXX (from 100 to 10000 m).}
\item{cabins_public_cumulative_exp_decayXXX} {Cumulative zone of influence of public resorts and mountain hotels
at each location, with exponential decay shape, and radii defined by XXX (from 100 to 10000 m).}
\item{trails_nearest_exp_decayXXX} {Zone of influence of the nearest tourist trail at each location,
with exponential decay shape, and radii defined by XXX (from 100 to 10000 m).}
\item{trails_cumulative_exp_decayXXX} {Cumulative zone of influence of tourist trails
at each location, with exponential decay shape, and radii defined by XXX (from 100 to 10000 m).}
}
}
\source{
Niebuhr, B. B., Van Moorter, B., Stien, A., Tveraa, T., Strand, O., Langeland, K.,
Sandström, P., Alam, M., Skarin, A., & Panzacchi, M. (2023). Estimating the cumulative impact
and zone of influence of anthropogenic features on biodiversity.
Methods in Ecology and Evolution. https://doi.org/10.1111/2041-210X.14133

Bakkestuen, V., Erikstad, L., & Halvorsen, R. (2008). Step-less models for regional
environmental variation in Norway. Journal of Biogeography, 35(10), 1906–1922.
https://doi.org/10.1111/j.1365-2699.2008.01941.x
}
\description{
Raster data with multiple layers representing predictor environmental variables
for the Hardangervidda wild reindeer areas in Norway, used for the vignettes and
the illustration example in Niebuhr et al. 2023.
Comprises multiple layers: land cover, four components of a principal component
analysis which represent bio-geo-climatic variation, and 28 layers representing
the cumulative zone of influence (ZOI) and the ZOI of the nearest feature for
private cottages and public tourist resources, with radii from 100 m to 20 km.
The raster has 500 m resolution.
}
\examples{
(f <- system.file("raster/hardanger_rast_predictors_500m.tif", package = "oneimpact"))
r <- terra::rast(f)
plot(r)

}
\seealso{
Data for RSF analysis, incorporating these predictors: \link{reindeer_rsf}
}
