% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_net_logit.R
\name{hypothesis_func}
\alias{hypothesis_func}
\title{Function to set penalties according to hypotheses}
\usage{
hypothesis_func(coefs, expectation = -1, phi_hyp = 1)
}
\arguments{
\item{coefs}{Vector of estimated coefficients through a Ridge regression.}

\item{expectation}{Expected/Hypothetical signal of the coefficient for a
given covariate. Should take value \code{hyp = -1} or \code{hyp = +1} depending
on the expected signal of the coefficient. If \code{expectation = 0}, no change is
made and the penalty if \code{penalty.factor = 1/abs(coef)**gamma}.}

\item{phi_hyp}{Additional penalty constant for the hypothesis-based penalties.
A value in the interval \link{1, Inf} where 1 is no additional penalty and
higher values correspond to higher penalties when}
}
\description{
Function to set penalties according to hypotheses
}
\examples{
# set coefficients
coefs <- c(-1, -0.5, -0.1, 0.8, 0.3, -0.1)
expected_sign <- -1
hypothesis_func(coefs)

x <- seq(-2, 2, length.out = 101)
plot(x, hypothesis_func(x), ylab = "Penalties")
plot(x, hypothesis_func(x, phi_hyp = 10), ylab = "Penalties")
plot(x, exp(x)*hypothesis_func(x, phi_hyp = 10), ylab = "Penalties")

}
\keyword{internal}
