% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_functions.R
\name{influence_functions}
\alias{influence_functions}
\alias{exp_decay}
\alias{threshold_decay}
\alias{step_decay}
\alias{bartlett_decay}
\alias{tent_decay}
\alias{linear_decay}
\alias{gaussian_decay}
\alias{half_norm_decay}
\title{Influence functions}
\usage{
exp_decay(
  x,
  zoi = NULL,
  exp_decay_parms = c(1, 0.01),
  zoi_decay_threshold = 0.05,
  origin = 0,
  half_life = NULL,
  zoi_hl_ratio = NULL,
  oneside = TRUE
)

threshold_decay(x, zoi, constant_influence = 1, origin = 0, oneside = TRUE)

step_decay(x, zoi, constant_influence = 1, origin = 0, oneside = TRUE)

bartlett_decay(x, zoi, intercept = 1, origin = 0, oneside = TRUE)

tent_decay(x, zoi, intercept = 1, origin = 0, oneside = TRUE)

linear_decay(x, zoi, intercept = 1, origin = 0, oneside = TRUE)

gaussian_decay(
  x,
  zoi = NULL,
  hnorm_decay_parms = c(1, 0.01),
  sigma = NULL,
  zoi_decay_threshold = 0.05,
  origin = 0,
  ...
)

half_norm_decay(
  x,
  zoi = NULL,
  hnorm_decay_parms = c(1, 0.01),
  sigma = NULL,
  zoi_decay_threshold = 0.05,
  origin = 0,
  ...
)
}
\description{
Decay influence functions, parameterized based on the zone of influence (ZoI).
}
\examples{
library(ggplot2)

# test influence functions

# exponential decay
exp_decay(10, zoi = 30)

f1 <- ggplot(data.frame(x = c(0, 30)), aes(x = x)) +
  stat_function(fun = exp_decay, args = list(zoi = 20)) +
  labs(x = "Distance", y = "Influence") +
  theme_bw()
f1

# exponential decay - two sides
f1_2 <- ggplot(data.frame(x = c(-30, 30)), aes(x = x)) +
  stat_function(fun = exp_decay, args = list(zoi = 20, oneside = FALSE)) +
  labs(x = "Distance", y = "Influence") +
  theme_bw()
f1_2

# threshold
threshold_decay(5, zoi = 10)
threshold_decay(10, zoi = 10)

f2 <- ggplot(data.frame(x = c(0, 30)), aes(x = x)) +
  stat_function(fun = threshold_decay, args = list(zoi = 20), linetype = 2) +
  labs(x = "Distance", y = "Influence") +
  theme_bw()
f2

# threshold - two sides
f2_2 <- ggplot(data.frame(x = c(-30, 50)), aes(x = x)) +
  stat_function(fun = threshold_decay, args = list(zoi = 20, oneside = FALSE), linetype = 2) +
  labs(x = "Distance", y = "Influence") +
  theme_bw()
f2_2

# linear, tent, or bartlett decay
bartlett_decay(5, zoi = 10)
bartlett_decay(8, zoi = 10)

f3 <- ggplot(data.frame(x = c(0, 30)), aes(x = x)) +
  stat_function(fun = bartlett_decay, args = list(zoi = 20), linetype = 3) +
  labs(x = "Distance", y = "Influence") +
  theme_bw()
f3

# linear, two sides
f3_3 <- ggplot(data.frame(x = c(-30, 30)), aes(x = x)) +
  stat_function(fun = bartlett_decay, args = list(zoi = 20, origin = 10, oneside = FALSE), linetype = 3) +
  labs(x = "Distance", y = "Influence") +
  theme_bw()
f3_3

# half normal
gaussian_decay(5, sigma = 6)

f4 <- ggplot(data.frame(x = c(0, 30)), aes(x = x)) +
  stat_function(fun = gaussian_decay, args = list(zoi = 20, zoi_decay_threshold = 0.05), linetype = 4) +
  labs(x = "Distance", y = "Influence") +
  geom_vline(xintercept = 20, linetype = 2, color = "darkgrey") +
  geom_hline(yintercept = 0.05, linetype = 2, color = "darkgrey") +
  theme_bw()
f4

# half normal - two sides
gaussian_decay(5, sigma = 6)

f4_2 <- ggplot(data.frame(x = c(-30, 30)), aes(x = x)) +
  stat_function(fun = gaussian_decay, args = list(zoi = 20, zoi_decay_threshold = 0.05), linetype = 4) +
  labs(x = "Distance", y = "Influence") +
  geom_vline(xintercept = c(-20, 20), linetype = 2, color = "darkgrey") +
  geom_hline(yintercept = 0.05, linetype = 2, color = "darkgrey") +
  theme_bw()
f4_2

# plot all
f1 +
  stat_function(fun = threshold_decay, args = list(zoi = 20), linetype = 2) +
  stat_function(fun = bartlett_decay, args = list(zoi = 20), linetype = 3) +
  stat_function(fun = gaussian_decay, args = list(zoi = 20, zoi_decay_threshold = 0.05), linetype = 4) +
  labs(x = "Distance", y = "Influence") +
  theme_bw()
}
