% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolation.R
\name{isolation}
\alias{isolation}
\alias{mean_isolation}
\title{Isolation and mean isolation of points in space}
\usage{
isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)

mean_isolation(x, n_rand = 100, ext = c(0, 1, 0, 1), lonlat = FALSE)
}
\arguments{
\item{x}{\verb{[data.frame]} \cr \code{data.frame} with (x,y) coordinates in
the columns.}

\item{n_rand}{\verb{[numeric(1)=100]} \cr Number of random points to be
created in space, to compute the distance to \code{x}.}

\item{ext}{\verb{[numeric(4)=c(0,1,0,1)]} \cr Extent of the space within which the
random positions should be created: \code{c(xmin, xmax, ymin, ymax)}.}

\item{lonlat}{\verb{[logical(1)=FALSE]} \cr Whether the distance between points
should be calculated in an WGS ellipsoid (\code{lonlat = TRUE}) or on a
plane (\code{lonlat = FALSE}). See \code{\link[raster:pointDistance]{raster::pointDistance()}} for more details.}
}
\value{
\code{isolation()} returns the distance from each random point to
the nearest neighbor point in \code{x}. \code{mean_isolation()} returns the average
nearest neighbor distance from all random positions to the points in
\code{x}.
}
\description{
Measures of isolation and mean isolation to a set of points in
space. \code{isolation()} creates random points in a landscape and calculates
the nearest neighbor distance from each of them to another set
of points passed as input, \code{x}. \code{mean_isolation()} calculates the
average isolation calculated through \code{isolation()}.
}
\details{
So far the function only works for a square landscape. In the future
we can implement that for polygons or rasters with masks or
null cells if necessary, in an approach similar to set_points_sample.
}
\examples{
pts <- set_points(n_features = 100, method = "random", centers = 1, width = 0.1)[[1]]
isolation(pts)
mean_isolation(pts)
}
