% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifit_multiscale.R
\name{multifit_multiscale}
\alias{multifit_multiscale}
\title{Evaluation of multiple covariate effects at multiple scales}
\usage{
multifit_multiscale(
  mod,
  mod_comparison_df,
  data,
  formula = NULL,
  weights = NULL,
  args = NULL,
  max_models = 10000,
  corr_check = TRUE,
  corr_criterion = c("usdm::vifcor", "usdm::vifstep")[1],
  corr_threshold = 0.7,
  progress = FALSE,
  verbose = FALSE,
  ...
)
}
\description{
Evaluation of multiple covariate effects at multiple scales
}
