% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifit.R
\name{multifit_single}
\alias{multifit_single}
\title{Evaluation of single covariate effects at multiple scales}
\source{
\url{https://github.com/phuais/multifit}
}
\usage{
multifit_single(
  mod,
  multief,
  data,
  formula = NULL,
  args = NULL,
  criterion = "AIC",
  site_id = NULL,
  signif = TRUE,
  alpha = 0.05,
  plot_est = FALSE,
  print_sum = FALSE,
  xlab = "Radius [km]",
  ylab = NULL,
  labels = NULL,
  type = "b",
  pch = c(1, 16)
)
}
\arguments{
\item{mod}{\verb{[character(1)]} \cr Type of model to be fit (e.g. "lm", "glm", "glmer").}

\item{multief}{\verb{[character]} \cr Vector containing the names of the columns of the input
\code{data.frame}, corresponding to the covariate, calculated at multiple zoi, radii, or scales,
whose fit will be assessed.}

\item{data}{\verb{[data.frame]} \cr \code{data.frame} containing the response variable and all model
covariates, including the covariate measured at multiple scales, whose fit will be assessed.}

\item{formula}{\verb{[formula]} \cr formula to be applied to each model...}
}
\description{
Evaluation of single covariate effects at multiple scales
}
\references{
Huais, PY. 2018. multifit: an R function for multi-scale analysis in landscape ecology. Landscape Ecology, 33: 1023. https://doi.org/10.1007/s10980-018-0657-5
}
\author{
Pablo Yair Huais <\href{mailto:phuais@gmail.com}{phuais@gmail.com}>
}
