% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifit_single_multivar.R
\name{multifit_single_multivar}
\alias{multifit_single_multivar}
\title{Evaluation of single covariate effects at multiple scales for multiple covariates}
\usage{
multifit_single_multivar(
  mod,
  covariates,
  data,
  formula = NULL,
  args = NULL,
  criterion = "AIC",
  print_best = 5,
  ...
)
}
\arguments{
\item{covariates}{\verb{[character]} \cr List of names of the covariates to be tested, or a pattern
within their names that differentiate among them.}
}
\description{
Evaluation of single covariate effects at multiple scales for multiple covariates
}
