% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_influence2d.R
\name{plot_influence2d}
\alias{plot_influence2d}
\title{Plot nearest and cumulative influence in 2 dimensions}
\usage{
plot_influence2d(
  points,
  zoi,
  fun = exp_decay,
  cumulative = FALSE,
  range_plot = c(0, 12),
  step = 0.01,
  na.rm = TRUE,
  return_df = FALSE,
  ...
)
}
\arguments{
\item{points}{Vector of values in the x axis where the infrastructure (or the origin
of the functions) are located}
}
\value{
a ggplot object with the nearest or cumulative influence plot.
If \code{return_df = TRUE}, it returns a list with the ggplot object and the \code{data.frame}
with values.
}
\description{
To add: options for normalization, prob etc.
}
\examples{
# exponential decay
points <- c(0, 4.5, 7, 7.5)
plot_influence2d(points, zoi = 3, fun = exp_decay, range_plot = c(0, 12), cumulative = FALSE)
plot_influence2d(points, zoi = 3, fun = exp_decay, range_plot = c(0, 12), cumulative = TRUE)

# threshold
plot_influence2d(points, zoi = 3, fun = threshold_decay, range_plot = c(0, 12), cumulative = F)
plot_influence2d(points, zoi = 3, fun = "step_decay", range_plot = c(0, 12), cumulative = T)

# linear
plot_influence2d(points, zoi = 3, fun = bartlett_decay, range_plot = c(0, 12), cumulative = F)
plot_influence2d(points, zoi = 3, fun = "linear_decay", range_plot = c(0, 12), cumulative = T, 
                 return_df = T)

# gaussian
plot_influence2d(points, zoi = 3, fun = gaussian_decay, range_plot = c(0, 12), cumulative = F)
plot_influence2d(points, zoi = 3, fun = "half_norm_decay", range_plot = c(0, 12), cumulative = T, 
                 return_df = T)
}
