% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_zoi1d.R
\name{plot_zoi1d}
\alias{plot_zoi1d}
\title{Plot the functions for the nearest and cumulative zone of influence in 1 dimension}
\usage{
plot_zoi1d(
  points,
  radius,
  fun = exp_decay,
  zoi_metric = c("nearest", "cumulative")[1],
  range_plot = c(0, 12),
  step = 0.01,
  na.rm = TRUE,
  return_df = FALSE,
  ...
)
}
\arguments{
\item{points}{\verb{[numeric]} \cr Vector of values in the x axis where the infrastructure are located,
in 1 dimension
(more broadly, the location of sources of disturbance or spatial variables, or the point of
origin of the ZOI functions).}

\item{radius}{\verb{[numeric(1)]} \cr Radius of the zone of influence (ZOI),
the distance at which the ZOI vanishes or goes below a given minimum limit value.
See \link{zoi_functions} for details.}

\item{fun}{\verb{[function]} \cr A decay function that represents the Zone of Influence (ZOI).
Different functions might represent different shapes for the decay of the ZOI.
See \link{zoi_functions} for some examples.}

\item{zoi_metric}{\verb{[character(1)="nearest"]\{"nearest", "cumulative"\}} \cr
Which metric of zone of influence should be plotted. If \code{"nearest"}(default), the ZOI of
the nearest feature is plotted. If \code{"cumulative"}, the cumulative ZOI is plotted.}

\item{range_plot}{\verb{[numeric(2)=c(0,12)]} \cr A vector \code{c(xmin, xmax)} with the x range of
the ZOI plot.}

\item{step}{\verb{[numeric(1)=0.01]} \cr Size of the step increment used to define the series
of x positions for which the ZOI is computed, within the x range defined by \code{range_plot}.}

\item{na.rm}{\verb{[logical(1)=TRUE]} \cr Should \code{NA} values be removed when computing
either the sum (for the cumulative ZOI) or the maximum (for the ZOI of the nearest
feature)?}

\item{return_df}{\verb{[logical(1)=FALSE]} \cr If TRUE, a \code{data.frame} with \code{x} values and their
corresponding ZOI values is returned, besides the plot.}

\item{...}{\verb{[any]} \cr Additional parameters passed to the ZOI decay function \code{fun}.}
}
\value{
A \code{ggplot} object with the nearest or cumulative influence plot.
If \code{return_df = TRUE}, it returns a list with the \code{ggplot} object and a \code{data.frame}  with
values for \code{x} (position in 1d space) and \code{y} (ZOI value).
}
\description{
Plots the functions to represent the zone of influence (ZOI) in 1 dimensional space, for illustration.
When there is more than one value for \code{points} (the location of infrastructure
or sources of disturbance), two metrics might be plotted: the ZOI of the nearest feature or the
cumulative ZOI. The ZOI of the nearest feature corresponds to the maximum ZOI value from all
infrastructure at each position. The cumulative ZOI corresponds to the sum of the ZOI
of all infrastructure at each position.
}
\details{
In practice, the \code{plot_zoi1d()} computes the ZOI value for each feature
whose locations in 1 dimension are defined by \code{points} and calculates the maximum
(ZOI of the nearest) or sum (cumulative ZOI) of all values. This is done
for a series of points in 1 dimensional space in the range \code{range_plot} (with
steps defined by \code{step}) and plotted.
}
\examples{
# one point, exponential decay
plot_zoi1d(0, radius = 3, zoi_limit = 0.05,
           fun = exp_decay, range_plot = c(0, 5), zoi_metric = "nearest")

# exponential decay
points <- c(0, 4.5, 7, 7.5)
plot_zoi1d(points, radius = 3, fun = exp_decay, range_plot = c(0, 12),
           zoi_metric = "nearest")
plot_zoi1d(points, radius = 3, fun = exp_decay, range_plot = c(0, 12),
           zoi_metric = "cumulative")

# threshold
plot_zoi1d(points, radius = 3, fun = threshold_decay, range_plot = c(0, 12),
           zoi_metric = "nearest")
plot_zoi1d(points, radius = 3, fun = "step_decay", range_plot = c(0, 12),
           zoi_metric = "cumulative")

# linear
plot_zoi1d(points, radius = 3, fun = bartlett_decay, range_plot = c(0, 12),
           zoi_metric = "nearest")
plot_zoi1d(points, radius = 3, fun = "linear_decay", range_plot = c(0, 12),
           zoi_metric = "cumulative")

# gaussian
plot_zoi1d(points, radius = 3, fun = gaussian_decay, range_plot = c(0, 12),
           zoi_metric = "nearest")
plot_zoi1d(points, radius = 3, fun = "half_norm_decay", range_plot = c(0, 12),
           zoi_metric = "cumulative", return_df = TRUE) # also return the data.frame
}
