% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_resamples.R
\name{pretty_seq}
\alias{pretty_seq}
\title{Create pretty, standardized sequences from numbers}
\usage{
pretty_seq(x, name = "Resample")
}
\arguments{
\item{x}{\verb{[vector,numeric]} \cr Vector of integers.}

\item{name}{\verb{[character(1)="Resample"]} \cr String to be
added to the numbers in the sequence.}
}
\description{
Create pretty, standardized sequences from numbers
}
\examples{
pretty_seq(1:20)

}
\keyword{internal}
