% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{reindeer_area.gpkg}
\alias{reindeer_area.gpkg}
\title{Reindeer area: a polygon vector data for the Setesdal Austhei reindeer herding area}
\format{
A geopackage file. Projected CRS: ETRS89 / UTM zone 33N.
}
\source{
Panzacchi, M., Van Moorter, B., Strand, O., Loe, L. E., & Reimers, E. (2015). Searching
for the fundamental niche using individual-based habitat selection modelling across
populations. Ecography, 38(7), 659–669. \url{https://doi.org/10.1111/ecog.01075}
}
\description{
Dataset containing the limits of the reindeer management area of Setesdal Austhei
in Southern Norway, as defined in Panzacchi et al. (2015). Note this area is slightly
larger than the boundaries used for reindeer management in Norway.
}
\examples{
(s <- system.file("vector/reindeer_area.gpkg", package = "oneimpact"))
sf::st_read(s)
# or
terra::vect(s)

}
