% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{reindeer_roads_public.gpkg}
\alias{reindeer_roads_public.gpkg}
\title{Public roads vector data for the reindeer area}
\format{
A geopackage file. Projected CRS: ETRS89 / UTM zone 33N. The vector
file presents the following columns:
\itemize{
\item{id:} {Line number, corresponding to the original dataset}
\item{name:} {Local name of the road}
\item{publ_priv:} {Whether the road is public or private}
\item{traffic_bin:} {Binary classification of the traffic on the road - high or low}
\item{name_area:} {Name of the reindeer management area where the road is located}
\item{traffic_bin:} {Value 1, to be used for rasterization purposes}
}
}
\source{
\url{https://kartkatalog.geonorge.no/metadata/elveg-20/77944f7e-3d75-4f6d-ae04-c528cc72e8f6}
}
\description{
Dataset containing the location of large, public roads in Southern
Norway, within the reindeer management area of Setesdal Austhei.
Retrieved from the Norwegian road dataset Elveg 1.0.
}
\examples{
(f <- system.file("vector/reindeer_roads_public.gpkg", package = "oneimpact"))
v <- sf::st_read(f)
plot(v[1])

}
