% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sample_area_cabins.gpkg}
\alias{sample_area_cabins.gpkg}
\title{Cabins vector data for the sample area}
\format{
A geopackage file. Projected CRS: ETRS89 / UTM zone 33N. The vector
file presents the following columns:
\itemize{
\item{cat:} {Line number, corresponding to the original dataset}
\item{buildtype:} {Type of building (code) in the original dataset}
\item{city:} {Code of the municipality where the cabin is located}
\item{value:} {Value 1, to be used for rasterization purposes}
}
}
\source{
\url{https://register.geonorge.no/det-offentlige-kartgrunnlaget/n50-kartdata/ea192681-d039-42ec-b1bc-f3ce04c189ac}
}
\description{
Dataset containing the location of tourist private cabins in Southern
Norway.
It corresponds to some specific building types (object_type = "Bygning",
byggtyp_nbr = c("161", "162", "163")) form the public N50 dataset.
The map was clipped for the sample area presented in the \code{oneimpact} package.
}
\examples{
(f <- system.file("vector/sample_area_cabins.gpkg", package = "oneimpact"))
sf::st_read(f)
# or
v <- terra::vect(f)
plot(v)

}
\seealso{
Maps for the sample area: \cr
Limits of sample area: \link{sample_area.gpkg} \cr
Cabins: \link{sample_area_cabins.tif}, \link{sample_area_cabins_count.tif} \cr
Roads: \link{sample_area_roads.gpkg}, \link{sample_area_roads.tif}
}
