% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_points_random.R
\name{set_points_random}
\alias{set_points_random}
\title{Simulate random points in 2D}
\usage{
set_points_random(n_features = 1000, extent_x = c(0, 1), extent_y = extent_x)
}
\arguments{
\item{n_features}{\verb{[integer(1)=1000]} \cr Total number of points to spread in space.}

\item{extent_x, entent_y}{\verb{[numeric vector(2)=c(0,1)]} \cr Vector representing the minimum and
maximum extent in x and y within which the points should be placed, in the format c(min,max).}
}
\value{
A \code{data.frame} with the (x,y) random coordinates.
}
\description{
This function simulates the coordinates of random points in
two dimensions by sampling (x,y) from a uniform distribution
between extremes. These extremes (the size of the landscape)
are defined by the \code{extent_x} and \code{extent_y} parameters.
}
\examples{
set_points_random(100)

}
