% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_sampling.R
\name{set_sampling_params}
\alias{set_sampling_params}
\title{Set sampling parameters}
\usage{
set_sampling_params(
  max_validation_size_blockH0 = 1000,
  max_number_blockH1 = 10,
  max_fit_size = 1000
)
}
\arguments{
\item{max_validation_size_blockH0}{= 1000,}

\item{max_number_blockH1}{= 10,}

\item{max_fit_size}{1000}
}
\value{
List of sampling parameters for the spatial stratified cross-validation,
with size of blocks of H0, H1, and H2.
}
\description{
This function sets the sampling parameters.
}
\examples{
set_sampling_params(1000, 10, 1000)

}
