% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_sampling.R
\name{set_validation_sampling}
\alias{set_validation_sampling}
\title{Setup blocks of points for fitting, calibration, and validation using spatial stratification}
\usage{
set_validation_sampling(spStrat, seed = 1234, sampling_params)
}
\arguments{
\item{spStrat}{A \code{data.frame} with spatially stratified levels in blocks H0 (population,
individual, or study area), H1 (level 1) and H2 (level 2). Each data point is associated
to a different set of blocks.}

\item{seed}{Seed for random sampling.}

\item{sampling_params}{Sampling parameters, the output from \link{set_sampling_params}.}
}
\value{
A vector with length equals to the number of positions, setting whether the
points are set to fitting (training), calibration (testing), and validation.
}
\description{
H0: validation
H1, H2
}
\examples{
library(terra)
library(amt)
data(reindeer)

# set spatially stratified blocks
spst <- spat_strat(reindeer, block_size = 10000, coords = c("x", "y"))
# set sampling parameters
sampling_params <- set_sampling_params(1000, 10, 1000)
# set train, test, and validation sets
set_validation_sampling(spst, seed = 1234, sampling_params)
# comment

}
