% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_strat.R
\name{spat_strat}
\alias{spat_strat}
\title{Preparing data for spatially stratified cross‐validation schemes}
\usage{
spat_strat(
  x,
  colH0 = NULL,
  colID = NULL,
  H1_as_H0 = FALSE,
  k = 4,
  block_size = 10000,
  buffer = 1000,
  coords = NULL,
  all_cols = FALSE,
  crs = "",
  plot_grid = TRUE,
  save_grid = c(NA_character_, "raster", "vector")[1]
)
}
\arguments{
\item{x}{\verb{[data.frame,sf,SpatVector]} \cr Vector of points to be spatially stratified, as a
\link{sf} or a \link[terra:SpatVector-class]{terra::SpatVector} object. If a \code{data.frame}, the columns corresponding to the (x,y) coordinates must be given in
\code{coords}.}

\item{colH0}{\verb{[numeric,character=NULL]} \cr Column number or name to define the ids of the H0 level
\itemize{
\item the one with ecological meaning, e.g. individuals, populations, or study areas, used for block cross-validating
the predictions of the fitted models. Default is \code{NULL}, in which case there is no block H0 defined.
}}

\item{H1_as_H0}{\verb{[logical(1)=FALSE]} \cr Whether the spatial blocks of level H1 should be used
as the block H0, in case no block H0 is provided (if \code{colH0 = NULL}). This parameter is ignored
if \code{colH0} is provided.}

\item{k}{\verb{[numeric(1)=4]} \cr Number of H2 blocks within each block H1. Should be 4, 9, 16, or some
number \code{k = x**2}, where x is an integer > 1. Default is \code{k = 4}.}

\item{block_size}{\verb{[numeric(1)=10000]} \cr Size (side of a square) of the blocks for H2 level,
map units (generally meters). The size of the H1 level blocks is defined as \code{sqrt(k)*block_size}.
Default is \code{block_size = 10000}}

\item{buffer}{\verb{[numeric(1)=1000]} \cr Buffer added around the points before creating the blocks,
to make sure all points are included in the samples. Default is \code{buffer = 1000}.}

\item{coords}{\verb{[character,vector]} \cr Vector with the names of the columns with the (x,y)
coordinates of the locations from the data set. Default is \code{NULL}, in which case \code{x}
should be a \code{sf} or a \code{terra::SpatVector} object. If \code{x} is a \code{data.frame}, \code{coords}
must be provided.}

\item{all_cols}{\verb{[logical=FALSE]} \cr If \code{TRUE}, and if \code{x} is a \code{data.frame},
the spatial strata blocks are appended as columns in the input data \code{x}.}

\item{crs}{\verb{[string=""]} \cr Coordinate reference system (CRS) of the observations,
if \code{x} is a \code{data.frame}.}

\item{plot_grid}{\verb{[logical=TRUE]} \cr if \code{TRUE} (default), the grid with spatial blocks and
observations is plotted.}

\item{save_grid}{\verb{[character=NA]\{NA, "raster", "vector"\}} \cr Should the grid which defines
the H1 and H2 blocks be saved? Not yet implemented.}

\item{col_id}{\verb{[numeric,character=NULL]} \cr Column number or name with the ID of the rows of the
data observations. In step-selection analysis, this should be the column showing the
number of the strata of each step. For resource selection analysis and environmental niche modeling,
this might be the row id, for instance.}
}
\value{
A \code{data.frame} with the blocks at hierarchical levels H0, H1, and H2 corresponding to
each of the observations in the input data set \code{x}.
}
\description{
This function sets up the variables used spatial stratification sampling to
be used in the bootstrapped modeling approach in \link{oneimpact}.
The function defines the hierarchical levels H0,
used for block cross-validation, as well as the levels H1 and H2, used for setting the
spatial structure of the sampling for model fitting and tuning.
Levels H1 and and H2 are organized in hierarchical spatial blocks. Blocks H1 correspond
to larger square blocks, each of which is subdivided into four, smaller H2 blocks
(with side equal \code{block_size}). H2 blocks are separated
spatially so that 3 of them are used for model fitting the the fourth is used for
model tuning.
The level H0 comes from the data set and should represent some other level for
block cross validation, such as population ID, animal ID, year or study area.
The block H0 is used for drawing validation sets, to allow a thorough evaluation
of the fitted models in all blocks H0.
}
\details{
The function returns a \code{data.frame} with the blocks for the hierarchical
levels H0, H1, and H2 for each observation in the input data set,
to be used to create samples for the bootstrapped modeling
approach using the \code{\link[=create_samples]{create_samples()}} function, in case spatial
stratified samples are desired.
}
\examples{
data(reindeer)
library(terra)
library(amt)

# no block H0, spatial blocks
spat_strat(reindeer, block_size = 5000, coords = c("x", "y"))

# with block H0
spst <- spat_strat(reindeer, coords = c("x", "y"), colH0 = "original_animal_id",
                   all_cols = TRUE)
# Visualize level H0 - individuals
spst_vect <- terra::vect(spst, geom = c("x", "y"))
terra::plot(spst_vect, "blockH0")
# Visualize level H1
terra::plot(spst_vect, "blockH1", type = "classes")
# Visualize level H2 for blockH1 numbers 6 to 10
terra::plot(spst_vect, col = grey(0.7))
terra::plot(spst_vect[spst$blockH1 == 6], "blockH2", type = "classes", add = TRUE) # only 6
terra::plot(spst_vect[spst$blockH1 \%in\% c(6,7,10,11)], "blockH2", type = "classes", add = TRUE) # 6-10
terra::plot(spst_vect[spst$blockH1 == 6], "blockH2", type = "classes") # zoom to 6

}
