% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_strat_fit_issf.R
\name{spat_strat_fit_issf}
\alias{spat_strat_fit_issf}
\title{Fit issf with spatial stratification and penalized regression}
\usage{
spat_strat_fit_issf(
  f,
  data,
  spStrat,
  sampling_params,
  i = 0,
  kernel_vars = c("step_length", "ta"),
  out_dir = NULL
)
}
\description{
Fit issf with spatial stratification and penalized regression
}
