% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{study_area.gpkg}
\alias{study_area.gpkg}
\title{Study area: a polygon vector data}
\format{
A geopackage file. Projected CRS: ETRS89 / UTM zone 33N.
}
\description{
Dataset containing the limits of an arbitrary study area in Southern Norway, used
for illustrative purposes.
}
\examples{
(s <- system.file("vector/study_area.gpkg", package = "oneimpact"))
sf::st_read(s)
# or
terra::vect(s)

}
