% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance.R
\name{variable_importance}
\alias{variable_importance}
\title{Computes variable importance}
\usage{
variable_importance(
  x,
  data,
  samples = NULL,
  type = c("drop", "permutation")[1],
  colH0 = NULL,
  variable_block = NULL,
  n_permutations = 100,
  order = c("desc", "asc", FALSE)[1],
  metric = NULL,
  plot = FALSE,
  ss = 1,
  remove_threshold = 0
)
}
\arguments{
\item{x}{\verb{[list]} \cr Bag of models, result of \code{\link[=bag_models]{bag_models()}}. It contains
multiple information from the models, such as formula, weights, coefficients,
and metric used to evaluate the models.}

\item{data}{\verb{[data.frame]} \cr Complete data set to which the models were applied.}

\item{samples}{\verb{[list]} \cr List of samples used to fit the models in the bag.
The list contains at least three elements: train, test,
and validate. Each elements might have several elements, each representing
the lines of \code{data} to be sampled for each resample. Typically, this is computed by
the function \code{\link[=create_resamples]{create_resamples()}}.}

\item{type}{\verb{[character(1)="drop"]\{"drop", "permutation"\}} \cr Type of computation
for variable importance. If \code{type = "drop"} (default), each variable is dropped from
the model at a time and the variation in model evaluation metric is computed.
If \code{type = "permutation"}, the observations of each variable are permutated and the
variation in the model evaluation metric is computed.}

\item{colH0}{\verb{[string(1)=NULL]} \cr String with the name of the column in \code{data}
representing the blockH0, in case we want the variable importance to be evaluated
for each block. Default is \code{NULL}, in case variable importance is assessed for all
the data.}

\item{variable_block}{\verb{[character,NULL]} \cr Optional grouping variable for computing
importance at the block level rather than per-variable. Default is \code{NULL}.}

\item{n_permutations}{\verb{[numeric(1)=100]} \cr Number of permutations, if
\code{type = "permutation"}.}

\item{order}{\verb{[character,logical(1)="desc"]\{"desc", "asc", FALSE\}} \cr Whether or
not to order the output variables according to descending (\code{order = "desc"}) or
ascending order of variable importance (\code{order = "asc"}). If \code{FALSE}, the variables
are shown in the same order as present in the bag of models, \code{x}.}

\item{metric}{\verb{[function,character=NULL]} \cr Metric function used to evaluate model performance.
If \code{NULL} (default), the metric stored in the bag \code{x} is used.}

\item{plot}{\verb{[logical(1)=FALSE]} \cr Should variable importance be plotted? Default
is \code{FALSE}.}

\item{ss}{\verb{[integer(1)=1]} \cr Index of the sample (resample) to use for validation when
\code{samples} is provided. Default is \code{1}.}

\item{remove_threshold}{\verb{[numeric(1)]} \cr Threshold for excluding variable with
little importance in the variable importance plot (i.e. only considered if \code{plot = TRUE}).
See more in \code{\link[=plot_importance]{plot_importance()}}.}
}
\value{
A data frame with variable names and their importance scores, optionally
ordered by importance and/or plotted.
}
\description{
Computation of variable importance from a bag of models. Variable importance can
be computed either by dropping variables or through permutation of the variables values
(parameter \code{type}), typically by evaluating the effects on the model evaluation metric
in the validation set(s). If \code{type = "drop"}, each variable is dropped from the model
at a time and the variation in model evaluation metric is computed. If \code{type = "permutation"},
The observations of each variable are permutated and the variation in model evaluation
metric is computed.
}
\seealso{
For plotting variable importance, see \code{\link[=plot_importance]{plot_importance()}}.
}
