% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weirdness.R
\name{weirdness}
\alias{weirdness}
\title{Computes ecological weirdness for a fitted model or it estimated  coefficients}
\usage{
weirdness(
  x,
  measure = c("coef_sign", "n_crosses", "response_auc"),
  wmean = TRUE,
  which_coef_sign = c("count", "sum")[1],
  expected_sign = -1,
  zero_coefficient_limit = 1e-08,
  response = c("mean", "mid")[1],
  radii = c(100, 250, 500, 1000, 2500, 5000, 10000),
  type = c("circle", "Gauss", "rectangle", "exp_decay", "bartlett", "threshold",
    "mfilter")[1],
  radius_max = 10000,
  ...
)
}
\arguments{
\item{x}{Bag.}

\item{measure}{\verb{[string(1)]\{""coef_sign", "n_crosses", "response_auc""\}} \cr Measure used
to quantify "weirdness" in the model or coefficients, based on the coeffcients and the response
plots for each type of covariate with zone of influence in a model.
It can be one or multiple of these options:
\itemize{
\item \code{"coef_sign"}: counts the number of coefficients whose signal is opposite to the ecologically expected signal;
\item \code{"n_crosses"}: counting the number of crosses in signal for the coefficients of the same covariate;
\item \code{"response_auc"}: computing the area under the response plot curve which is in the unexpected
direction.
}}

\item{expected_sign}{\verb{[numeric(1)=-1]} \cr Expected sign of the coefficient. Either -1 (negative),
+1 (positive), or 0 (no effect).}

\item{zero_coefficient_limit}{\verb{[numeric(1)=1e8]} \cr Value above which an estimated coefficient is considered
non-zero. Default is 1e-8. Useful for comparing coefficients which are expected to be zero (i.e. to have no effect).}

\item{which_coef}{\cr Which measure to use for the coefficients, when \code{measure = "coef_sign"}. If \code{count} (default),
only the sign matterns and we count the number of coefficients with unexpected sign.
If \code{sum}, we count the sum of the (standardized) coefficients, to also account for their magnitude.}
}
\description{
Computes ecological weirdness for a fitted model or it estimated  coefficients
}
