% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zoi_formula.R
\name{add_zoi_formula}
\alias{add_zoi_formula}
\title{Adds ZOI radii to formula}
\usage{
add_zoi_formula(f, zoi_radius, pattern = "XXX")
}
\description{
Adds multiple radii for zone of influence variable to a formula, to be fitted in
a statistical model.
}
\examples{
f <- case_ ~ strata(step_id_) + sl_*startpt_roadsXXX + sl_*startpt_cabinsXXX
add_zoi_formula(f, zoi_radius = c(1000, 2000, 3000), pattern = "XXX")

}
