% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zoi_layers.R
\name{add_zoi_layers}
\alias{add_zoi_layers}
\title{Create ZOI layer names as strings for data annotation}
\usage{
add_zoi_layers(layers, zoi_radius, name = NULL, pattern = NULL)
}
\description{
This function uses a vector of variable names and a vector of ZOI radii to create
a \code{data.frame} with all combinations of variable names and zoi radii, to ease
listing or accessing them from GRASS GIS or for setting up a model formula.
}
\examples{
# attaching zoi_radius to the end
name <- c("houses", "roads", "railways")
radii <- seq(100, 500, by = 100)
add_zoi_layers(name, radii)

# replacing a pattern by zoi_radius
themes <- c("houses_XXX", "private_cabins_XXX", "roads_XXX", "powerlines_XXX", "railways_XXX")
maps <- c("houses_XXX@my_mapset", "private_cabins_XXX@my_mapset", "roads_summer_XXX@my_mapset",
         "powerlines_XXX@my_mapset", "railway_XXX@my_mapset")

add_zoi_layers(layers = maps, zoi_radius = c(100, 500, 1000, 5000, 10000),
               name = themes, pattern = "XXX")

}
