% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_load_models.R
\name{bag_load_models}
\alias{bag_load_models}
\title{Load a series of files output of fit_net_clogit models and put them on a bag}
\usage{
bag_load_models(
  f,
  data,
  samples,
  load_models_path = ".",
  load_models_pattern = NULL,
  names_from_file = FALSE,
  name_from_file_pattern = "Resample",
  metric = c(conditionalBoyce, conditionalSomersD, conditionalAUC)[[1]],
  standardize = c("internal", "external", FALSE)[1],
  method = c("Lasso", "Ridge", "AdaptiveLasso", "DecayAdaptiveLasso", "ElasticNet")[1],
  verbose = FALSE
)
}
\arguments{
\item{f}{\verb{[formula]} \cr Formula of the models fitted, with all possible candidate terms.}

\item{data}{\verb{[data.frame,tibble]} \cr Complete data set analyzed.}

\item{samples}{\verb{[list]} \cr List of samples with at least three elements: train, test,
and validate. Each elements might have several elements, each representing
the lines of \code{data} to be sampled for each resample. Typically, this is computed by
the function \code{\link[=create_resamples]{create_resamples()}}.}

\item{load_models_path}{\verb{[string="."]} \cr Path to the folder where the files
are saved.}

\item{load_models_pattern}{\verb{[string="."]} \cr Pattern common to the file names,
to be used in \code{grep()} to find the files within the folder
\code{load_models_path}. It should be the pattern present in the argument
\code{out_dir_file} in \code{\link[=fit_net_clogit]{fit_net_clogit()}} or \code{\link[=bag_fit_net_clogit]{bag_fit_net_clogit()}}
functions.}

\item{names_from_file}{\verb{[logical(1)=TRUE]} \cr If \code{FALSE} (default), the names of
the resamples are taken from the \code{samples} parameter. If \code{TRUE}, they are taken
from the files names, instead. In this case, the string defined by the parameter
\code{name_from_file_pattern} is used to identify the number of each resample.}

\item{name_from_file_pattern}{\verb{[string="Resample"]} \cr String used to separate the
file name and identify the number of the resample of each model in the bag, when
they are read from files.

The parameters metric, standardize, and method should be same ones used to
fit the bag of models.}
}
\description{
Load a series of files output of fit_net_clogit models and put them on a bag
}
