% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_models.R
\name{bag_models}
\alias{bag_models}
\title{Summary of a bag of models}
\usage{
bag_models(
  fitted,
  data,
  score2weight = NULL,
  weights_col = c("validation_score", "habitat_validation_score")[1],
  score_threshold = 0.7,
  weights_function = NULL,
  out_dir
)
}
\arguments{
\item{data}{Input data (whole dataset)}

\item{score2weight}{Function to set validation scores into weights, with two arguments:
x, the result of one model of the bag, and col, the column to be used for setting the
scores. See the argument \code{weights_col}.}

\item{weights_col}{Column to use for scores. One of \code{"validation_score"} or
\code{"habitat_validation_score"}.}
}
\description{
Summary of a bag of models
}
