% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_models.R
\name{bag_models}
\alias{bag_models}
\alias{w_strech_maxmin_squared}
\alias{w_strech_max_squared}
\alias{score2weight_mean}
\alias{score2weight_min_mean}
\alias{score2weight_invmean}
\alias{score2weight_min_invmean}
\title{Summary of a bag of models}
\usage{
bag_models(
  fitted,
  data,
  score2weight = score2weight_min_invmean,
  weights_col = c("validation_score", "habitat_validation_score")[1],
  score_threshold = 0.7,
  weights_function = NULL,
  out_dir
)

w_strech_maxmin_squared(x)

w_strech_max_squared(x)

score2weight_mean(x, col = "validation_score", score_threshold = 0.7)

score2weight_min_mean(x, col = "validation_score", score_threshold = 0.7)

score2weight_invmean(x, col = "validation_score", score_threshold = 0.7)

score2weight_min_invmean(x, col = "validation_score", score_threshold = 0.7)
}
\arguments{
\item{data}{Input data (whole dataset)}

\item{score2weight}{Function to set validation scores into weights, with two arguments:
x, the result of one model of the bag, and col, the column to be used for setting the
scores. See the argument \code{weights_col}.}

\item{weights_col}{Column to use for scores. One of \code{"validation_score"} or
\code{"habitat_validation_score"}.}
}
\description{
\subsection{Add option to read the models from a set of rds/rda files.}{
}
}
