% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_predict.R
\name{bag_predict}
\alias{bag_predict}
\title{Prediction of a bag of models to new data}
\usage{
bag_predict(
  x,
  newdata,
  type = c("linear", "exponential", "exp", "logit", "cloglog")[1],
  wMean = T,
  wq_probs = NULL,
  include = "all"
)
}
\arguments{
\item{x}{\verb{[bag,list]} \cr  A bag of models, resulting from a call to \code{\link[=bag_models]{bag_models()}}.}

\item{newdata}{\cr New data set to be used for prediction.}

\item{inclue}{vector with 1s and 0s about which variables should be included?}
}
\description{
Prediction of a bag of models to new data
}
