% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_predict_spat.R
\name{bag_predict_spat}
\alias{bag_predict_spat}
\title{Predict bag of models in space}
\usage{
bag_predict_spat(
  bag,
  data,
  input_type = c("df", "rast")[1],
  output_type = c("df", "rast")[2],
  baseline_max = NULL,
  gridalign = TRUE,
  gid = "gid",
  coords = c("x33", "y33"),
  crs = NULL,
  plotit = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{gid}{\verb{[string(1)="gid"]} \cr String with the name of the "gid" or point ID
column in \code{data}.}
}
\description{
Predict bag of models in space
}
