% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dens.R
\name{calc_dens}
\alias{calc_dens}
\title{Calculate density of features at multiple scales}
\usage{
calc_dens(
  points,
  type_density = c("Gauss", "circle", "rectangle")[1],
  scale = 100,
  extent_x_cut = bbox(points)[1, ],
  extent_y_cut = bbox(points)[2, ],
  plotit = FALSE,
  ...
)
}
\arguments{
\item{points}{\verb{[RasterLayer]} \cr Raster representing locations of features, with 1 where the features
are localted and NA elsewhere.}

\item{type_density}{\verb{[character(1)="Gauss"]\{"Gauss", "circle", "rectangle"\}} \cr
Type of filter used to calculate density. See description for details.}

\item{scale}{\verb{[numeric(1)=100]} \cr Scale of the neighborhood analysis, used to calculate densities.
It can be a single value of a vector of values, in which case several density maps (for each scale)
are created.}

\item{extent_x_cut, entent_y_cut}{\verb{[numeric vector(2)=c(0,1)]} \cr Vector representing the minimum and
maximum extent in x and y for the final output, in the format c(min,max). It is intended to keep only
a region of interest but consider the surroundings when calculating densities. The default is to
keep the same extent of the input raster.}

\item{plotit}{\verb{[logical(1)=FALSE]} \cr Should the outputs be plotted along the calculation?}
}
\value{
A RasterBrick with the density of features for all scales selected.
}
\description{
This function takes in a raster with locations of infrastructure and calculates
a raster (or set of rasters, in case there is more the one value for \code{scale})
representing the density of features in space (through a spatial filter/neighborhood analysis).
The neighborhood analysis is done with the \code{\link[raster:focal]{raster::focal()}} function.
}
\details{
The neighborhood analysis can be done with different methods. The default is a Gaussian filter
(\code{type_density = "Gauss"}), in which case scale corresponds to the sigma paramater of the Gaussian
filter. If \code{type_density = "circle"} or \code{type_density = "rectangle"}, the scale corresponds to the
radius of the circle or width of the rectangle, respectively. See \code{\link[raster:focalWeight]{raster::focalWeight()}} for more
details.

TO IMPROVE1: implement with \code{terra}.

TO IMPROVE2: do the same in communication with GRASS GIS.
}
\examples{
library(mobsim)
library(raster)

set.seed(1234)

# set points
ext <- 30000
wd <- ext/20
pts <- set_points(n_features = 1000, centers = 1,
                  width = wd, res = 100,
                  extent_x = c(0, ext), extent_y = c(0, ext),
                  buffer_around = 10000)
plot(pts$pts)
plot(pts$rast)

# calculate distance and densities considering only the initial extent
scales <- c(250, 500, 1000, 2500, 5000)/2
dens <- calc_dens(pts$rast, type_density = "Gauss", scale = scales,
                  extent_x_cut = c(0, ext), extent_y_cut = c(0, ext))
plot(dens)

# calculate log_dist (the rest is equal)
dens_circle <- calc_dens(pts$rast, type_density = "circle", scale = 2*scales,
                         extent_x_cut = c(0, ext), extent_y_cut = c(0, ext))
plot(dens_circle)
}
