% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_zoi_sql.R
\name{calc_zoi_sql}
\alias{calc_zoi_sql}
\title{Compute Zone of Influence for points and annotate them to points using SQL}
\usage{
calc_zoi_sql(
  input_points,
  infrastructure_layer,
  radius = 100,
  type = c("circle", "Gauss", "exp_decay", "bartlett", "threshold")[1],
  zoi_metric = c("cumulative", "nearest")[1],
  input_id = "id",
  input_geom = "geom",
  infra_geom = "geom",
  output_type = c("cumulative_zoi", "density")[1],
  zoi_limit = 0.05,
  output_column_name = paste0(infrastructure_layer, "_", output_type, "_bartlett",
    radius),
  output_table = NULL,
  condition = "",
  limit = 1e+15,
  verbose = FALSE
)
}
\arguments{
\item{input_points}{\verb{[character]} \cr Name of the input table of points to be annotated (within the connection \code{con},
a database), in the format \code{table_name} or \code{schema_name.table_name}.}

\item{infrastructure_layer}{\verb{[character]} \cr Name of the infrastructure covariate table for which
the zone of influence will be computed (within the connection \code{con}, a database),
in the format \code{table_name} or \code{schema_name.table_name}.}

\item{radius}{\verb{[numeric=100]} \cr Radius or scale of zone of influence, used to calculate the
cumulative ZOI and density. The radius represent the distance at which the ZOI vanishes or
goes below a given minimum limit value \code{zoi_limit}. ### CHECK THAT, so far only bartlett}

\item{input_id}{\verb{[character]} \cr Name of a ID column from the \code{input_points} table.}

\item{input_geom}{\verb{[character]} \cr Name of the geometry column from the \code{input_points} table.}

\item{infra_geom}{\verb{[character]} \cr Name of the geometry column from the \code{infrastructure_layer} table.}

\item{output_type}{\verb{[character]} \cr}

\item{output_column_name}{\verb{[character]} \cr}
}
\description{
This function computes the zone of influence (ZOI) for a set of points
(typically a type of indrastructure, defined by the \code{infrastructure_layer} parameter)
and extracts their values at given location (defined by the layer \code{input_points}),
using a SQL query into a database (e.g. PostgreSQL or duckdb). Differently from
the \code{calc_zoi} functions, it does not compute the ZOI for the whole study area,
but only for the positions where the input points are located. For optimal use,
both \code{input_points} and \code{infrastructure_layer} should be spatially indexed, so
only closeby points are evaluated and the query is completed fast.
}
\details{
Need to check functions (gaussian, exp, linear) and check if this is correct
}
