% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{concordance_indices}
\alias{concordance_indices}
\alias{somersD}
\alias{AUC}
\title{Computes concordance indices for model evaluation}
\usage{
somersD(x, errors = TRUE, warnings = TRUE)

AUC(x, errors = TRUE, warnings = TRUE)
}
\description{
These functions compute different concordance indices.
}
\references{
https://cran.r-project.org/web/packages/survival/vignettes/concordance.pdf
}
\seealso{
\link{conditionalBoyce}
}
