% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{conditionalBoyce}
\alias{conditionalBoyce}
\title{Computes the conditional Boyce index for model evaluation}
\usage{
conditionalBoyce(
  x,
  method = c("pearson", "kendall", "spearman")[1],
  plotit = FALSE,
  errors = TRUE,
  warnings = TRUE
)
}
\arguments{
\item{x}{A data.frame with three columns: x, the predicted values; y, the case variable
(use vs. available, 1/0); and strat, the stratum.}
}
\description{
Computes the conditional Boyce index for model evaluation
}
\details{
\code{AUC} is the implementation of the computation of the Area Under the Curva as related
to the Sommers'D index,
as described in https://cran.r-project.org/web/packages/survival/vignettes/concordance.pdf.
\code{proc_AUC} uses the \code{\link[pROC:auc]{pROC::auc()}} function.
}
