% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_resamples.R
\name{create_resamples}
\alias{create_resamples}
\title{Create samples for fitting, calibrating, and validating models}
\usage{
create_resamples(
  y,
  times = 10,
  p = c(0.4, 0.2, 0.2),
  list = TRUE,
  replace = TRUE
)
}
\arguments{
\item{y}{a vector of outcomes.}

\item{times}{the number of partitions or samples to create}

\item{p}{a 3 element vector with the percentage of data that goes to training, validation, and test}

\item{list}{logical - should the results be in lists (TRUE) or matrices with the number of rows equal to
floor(p * length(y)) and times columns.}

\item{replace}{whether to perform the bootstrap sampling with or without replacement (Default = TRUE).}
}
\description{
The function creates samples of the data to be included for three different
model blocks: train (fitting), validating, and testing models.
By default, samples are created through bootstrapping, i.e. with replacement. This means
data observations can be repeated within a given sample block, but observations included
in one block are necessarily excluded from the other blocks (e.g. observations selected
for test will be absent from train and validation blocks).
}
\examples{
y <- runif(200)
samples <- create_resamples(y, p = c(0.4, 0.2, 0.2), times = 5)
samples

}
