% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_net_clogit.R
\name{fit_net_clogit}
\alias{fit_net_clogit}
\alias{fit_net_ssf}
\alias{fit_net_issf}
\title{Fits a conditional logistic regression/SSF/iSSF using glmnet in a train-validate-test setup}
\usage{
fit_net_clogit(
  f,
  data,
  samples,
  i = 1,
  kernel_vars = c("step_length", "ta"),
  metric = c(conditionalBoyce, somersD, AUC)[[1]],
  out_dir_file = NULL,
  na.action = "na.pass",
  ...
)

fit_net_ssf(
  f,
  data,
  samples,
  i = 1,
  kernel_vars = c("step_length", "ta"),
  metric = c(conditionalBoyce, somersD, AUC)[[1]],
  out_dir_file = NULL,
  na.action = "na.pass",
  ...
)

fit_net_issf(
  f,
  data,
  samples,
  i = 1,
  kernel_vars = c("step_length", "ta"),
  metric = c(conditionalBoyce, somersD, AUC)[[1]],
  out_dir_file = NULL,
  na.action = "na.pass",
  ...
)
}
\arguments{
\item{...}{Options for net_clogit and glmnet}
}
\description{
Fits a conditional logistic regression/SSF/iSSF using glmnet in a train-validate-test setup
}
