% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_net_logit.R
\name{fit_net_logit}
\alias{fit_net_logit}
\alias{fit_net_rsf}
\title{Fit logistic regression/RSF with penalized regression in a train-validate-test setup}
\usage{
fit_net_logit(
  f,
  data,
  samples,
  i = 1,
  metric = c(conditionalBoyce, somersD, AUC, proc_AUC)[[1]],
  out_dir_file = NULL,
  na.action = "na.pass",
  ...
)

fit_net_rsf(
  f,
  data,
  samples,
  i = 1,
  metric = c(conditionalBoyce, somersD, AUC, proc_AUC)[[1]],
  out_dir_file = NULL,
  na.action = "na.pass",
  ...
)
}
\arguments{
\item{...}{Options for net_logit and glmnet}
}
\description{
Fit logistic regression/RSF with penalized regression in a train-validate-test setup
}
