% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_net_logit.R
\name{fit_net_logit}
\alias{fit_net_logit}
\alias{fit_net_rsf}
\title{Fit logistic regression/RSF with penalized regression in a train-validate-test setup}
\usage{
fit_net_logit(
  f,
  data,
  samples,
  i = 1,
  metric = c(AUC, conditionalBoyce, conditionalSomersD, conditionalAUC)[[1]],
  method = c("Lasso", "Ridge", "AdaptiveLasso", "DistanceDecay-AdaptiveLasso",
    "DD-AdaptiveLasso", "OneZOI-AdaptiveLasso", "OZ-AdaptiveLasso",
    "Grouped-AdaptiveLasso", "G-AdaptiveLasso", "HypothesisDriven-AdaptiveLasso",
    "HD-AdaptiveLasso", "ElasticNet")[1],
  alpha = NULL,
  penalty.factor = NULL,
  gamma = 1,
  standardize = c("internal", FALSE)[1],
  predictor_table = NULL,
  function_lasso_decay = c(log, function(x) x/1000)[[1]],
  value_lasso_decay = 1,
  factor_hypothesis = 1,
  factor_grouped_lasso = 1,
  na.action = "na.pass",
  out_dir_file = NULL,
  ...
)

fit_net_rsf(
  f,
  data,
  samples,
  i = 1,
  metric = c(AUC, conditionalBoyce, conditionalSomersD, conditionalAUC)[[1]],
  method = c("Lasso", "Ridge", "AdaptiveLasso", "DistanceDecay-AdaptiveLasso",
    "DD-AdaptiveLasso", "OneZOI-AdaptiveLasso", "OZ-AdaptiveLasso",
    "Grouped-AdaptiveLasso", "G-AdaptiveLasso", "HypothesisDriven-AdaptiveLasso",
    "HD-AdaptiveLasso", "ElasticNet")[1],
  alpha = NULL,
  penalty.factor = NULL,
  gamma = 1,
  standardize = c("internal", FALSE)[1],
  predictor_table = NULL,
  function_lasso_decay = c(log, function(x) x/1000)[[1]],
  value_lasso_decay = 1,
  factor_hypothesis = 1,
  factor_grouped_lasso = 1,
  na.action = "na.pass",
  out_dir_file = NULL,
  ...
)
}
\arguments{
\item{f}{\verb{[formula]} \cr Formula of the model to be fitted, with all possible candidate terms.}

\item{data}{\verb{[data.frame,tibble]} \cr Complete data set to be analyzed.}

\item{samples}{\verb{[list]} \cr List of samples with at least three elements: train, test,
and validate. Each elements might have several elements, each representing
the lines of \code{data} to be sampled for each resample. Typically, this is computed by
the function \code{\link[=create_resamples]{create_resamples()}}.}

\item{metric}{\verb{[function]\{AUC, conditionalBoyce, conditionalSomersD, conditionalAUC\}} \cr Function
representing the metric to evaluate goodness-of-fit. One of AUC (Default), conditionalBoyce,
conditionalSomersD, and conditionalAUC. A user-defined function might be provided, with a condition that
it must be maximized to find the best fit model.}

\item{method}{\verb{[character="Lasso"]} \cr The penalized regression method used for fitting
each model. Default is \code{method = "Lasso"}, but it could be \code{method = "Ridge"} or different
flavors of \code{"AdaptiveLasso"} (see details below).}

\item{gamma}{\verb{[numeric(1)=1]\{(0.5, 1, 2)\}} \cr Gamma is the exponent for defining the vector of
penalty weights when \verb{method = "AdaptiveLasso}. This means that the penalties are defined as
\code{penalty.factor = 1/(coef_ridge^gamma)}, where \code{coef_ridge} are the coefficients of a Ridge regression.
Default is \code{gamma = 1}, but values of 0.5 or 2 could also be tried, as suggested by the
authors (Zou et al 2006).}

\item{standardize}{\verb{[logical(1)=TRUE]} \cr Logical flag for predictor variable standardization,
prior to fitting the model sequence. The coefficients are always returned on the original scale.
Default is standardize=TRUE. If variables are in the same units already, you might not wish to
standardize them.}

\item{out_dir_file}{\verb{[character(1)=NULL]} \cr String with the prefix of the file name (and
the folder) where the result of each model will be saved. E.g. if \code{out_dir_file = "output/test_"},
the models will be saved as RDS files names "test_i1.rds", "test_i2.rds", etc, within the
folder "output".}

\item{...}{Options for net_logit and glmnet}
}
\description{
By default, \code{\link[=fit_net_logit]{fit_net_logit()}} does not standardize predictor variables. If you want numeric variables
to be standardized, you can either use \verb{[bag_git_net_logit()]} with parameter \code{standardize = TRUE}
or provide an already standardized data set as input.
}
\references{
Zou, H., 2006. The Adaptive Lasso and Its Oracle Properties. Journal of the American Statistical Association 101, 1418–1429. https://doi.org/10.1198/016214506000000735
}
