% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hardanger_trails.gpkg}
\alias{hardanger_trails.gpkg}
\title{Tourist trails vector data for the Hardangervidda wild reindeer area in Norway}
\format{
A geopackage file. Projected CRS: ETRS89 / UTM zone 33N. The vector
file presents the following columns:
\itemize{
\item{gid:} {Line number/identifier}
\item{area:} {Name of the wild reindeer area, if within one}
\item{traffic_bin:} {Binary classification of the tourist traffic on the trail - high or low}
\item{pseudotui:} {Continuous variable representing relative number of tourists per trail.}
\item{value:} {Value 1, to be used for rasterization purposes}
}
}
\source{
\url{https://register.geonorge.no/det-offentlige-kartgrunnlaget/n50-kartdata/ea192681-d039-42ec-b1bc-f3ce04c189ac}
}
\description{
Dataset containing the location of tourist trails in the
surroundings of the Hardangervidda wild reindeer area.
Retrieved from the public N50 dataset.
}
\examples{
(f <- system.file("vector/hardanger_trails.gpkg", package = "oneimpact"))
v <- sf::st_read(f)
plot(v[1])

}
