% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weirdness.R
\name{inflection}
\alias{inflection}
\title{Find inflection points in a curve}
\usage{
inflection(x)
}
\arguments{
\item{`x`}{\verb{[vector,numeric]} \cr Vector of \code{y} values, i.e., the response variable of
a curve.}
}
\description{
Helper function to find inflection points in a curve, using the response axis values.
}
\keyword{internal}
