% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_prediction.R
\name{kernel_prediction}
\alias{kernel_prediction}
\title{Prediction based only on step length terms}
\usage{
kernel_prediction(f, data, kernel_vars = c("step_length", "ta"), coefs)
}
\description{
Prediction based only on step length terms
}
