% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_issf.R
\name{net_clogit}
\alias{net_clogit}
\alias{net_ssf}
\alias{net_issf}
\title{Fits a conditional logistic regression/SSF/iSSF using glmnet}
\usage{
net_clogit(
  f,
  data,
  alpha = 1,
  type.measure = "deviance",
  na.action = "na.pass",
  ...
)

net_ssf(
  f,
  data,
  alpha = 1,
  type.measure = "deviance",
  na.action = "na.pass",
  ...
)

net_issf(
  f,
  data,
  alpha = 1,
  type.measure = "deviance",
  na.action = "na.pass",
  ...
)
}
\arguments{
\item{f}{formula}

\item{data}{data}
}
\description{
Fits a conditional logistic regression/SSF/iSSF using glmnet

Function with similar name

Function with similar name
}
