% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_logit.R
\name{net_logit}
\alias{net_logit}
\alias{net_rsf}
\title{Fits a logistic regression/RSF using glmnet}
\usage{
net_logit(
  f,
  data,
  alpha = 1,
  type.measure = "deviance",
  standardize = TRUE,
  na.action = "na.pass",
  ...
)

net_rsf(
  f,
  data,
  alpha = 1,
  type.measure = "deviance",
  standardize = TRUE,
  na.action = "na.pass",
  ...
)
}
\arguments{
\item{f}{formula}

\item{data}{data}

\item{alpha}{Default is L1-regularization (Lasso regression), with \code{alpha = 1}.
L2-regularization (Ridge regression) is done with \code{alpha = 0}, and elastic-net regression
is performed for any \code{alpha} value between \code{0} and \code{1}. For more details, see the
\code{\link[glmnet:glmnet]{glmnet::glmnet()}} documentation.}

\item{na.action}{Default is \code{"na.pass"}, i.e. rows with NAs are not automatically
removed from the \code{model.matrix} used for fitting.

Check option parallel = TRUE from glmnet.}
}
\description{
Fits a logistic regression/RSF using glmnet

Function with similar name
}
