% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_logit.R
\name{net_logit}
\alias{net_logit}
\alias{net_rsf}
\title{Fits a logistic regression/RSF using glmnet}
\usage{
net_logit(
  f,
  data,
  alpha = 1,
  penalty.factor = NULL,
  type.measure = "deviance",
  standardize = TRUE,
  na.action = "na.pass",
  func = c("glmnet", "cv.glmnet")[1],
  ...
)

net_rsf(
  f,
  data,
  alpha = 1,
  penalty.factor = NULL,
  type.measure = "deviance",
  standardize = TRUE,
  na.action = "na.pass",
  func = c("glmnet", "cv.glmnet")[1],
  ...
)
}
\arguments{
\item{f}{\verb{[formula]} \cr Formula of the model to be fitted, with all possible candidate terms.}

\item{data}{\verb{[data.frame,tibble]} \cr Complete data set to be analyzed.#' @param alpha Default is L1-regularization (Lasso regression), with \code{alpha = 1}.
L2-regularization (Ridge regression) is done with \code{alpha = 0}, and elastic-net regression
is performed for any \code{alpha} value between \code{0} and \code{1}. For more details, see the
\code{\link[glmnet:glmnet]{glmnet::glmnet()}} documentation. For Adaptive and Decay Adaptive Lasso, keep \code{alpha = 1}.}

\item{penalty.factor}{\verb{[numeric,vector=NULL]} \cr Vector of penalty factors to be used for Adaptive Lasso
fitting. The vector might have the same length as the the number of columns given by the model matrix,
\code{model.matrix(f, data)}. Default is \code{NULL}, in case the same penalty is applied to all variables.}

\item{type.measure}{\verb{[character(1)="deviance"]} \cr Type of measure to evaluate the model internally
in \code{\link[glmnet:glmnet]{glmnet::glmnet()}}. For logistic and conditional logistic regression, it is by default \code{"deviance"}.}

\item{na.action}{\verb{[character(1)="na.pass"]} \cr Default is \code{"na.pass"}, i.e. rows with NAs are not automatically
removed from the \code{model.matrix} used for fitting.}

\item{func}{\verb{[character(1)="glmnet"]\{"glmnet", "cv.glmnet"\}} \cr The function to be used for
fitting. Default is \code{\link[glmnet:glmnet]{glmnet::glmnet()}}. The second option is \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}} which
already performs the cross-validation and might include the variable selection/callibration
within.

Check option parallel = TRUE from glmnet.}
}
\description{
Fits a logistic regression/RSF using glmnet

Function with similar name
}
