% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coef.R
\name{plot_coef}
\alias{plot_coef}
\title{Plot the coefficients of bags of models}
\usage{
plot_coef(
  bag,
  terms = "all",
  models = 1:bag$n_above_threshold,
  weighted = TRUE,
  remove_weight_zero = TRUE,
  what = c("all_models", "average")[1],
  plot_type = c("bars", "points", "histogram")[1],
  remove_low = -1,
  remove_high = Inf,
  std = c(FALSE, "std", "unstd")[1],
  data = NULL,
  order_zoi_radius = FALSE,
  show_legend = FALSE
)
}
\description{
Options - one term, multiple bars side by side
Colors for negative and positive
}
