% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance.R
\name{plot_importance}
\alias{plot_importance}
\title{Plot variable importance}
\usage{
plot_importance(
  importance,
  remove_threshold = 0,
  normalize = TRUE,
  summ_stat = c(FALSE, "mean")[1]
)
}
\arguments{
\item{importance}{\verb{[numeric]} \cr Numeric vector showing the importance score of
each variable, resulting from \code{\link[=variable_importance]{variable_importance()}}.}

\item{remove_threshold}{\verb{[numeric(1)=0]} \cr Only importance scores above this
level will be plotted. Default is 0, in case which all null variables will be
removed. To plot all variables, set remove_threshold as -1 or any other negative value.}

\item{normalize}{\verb{[logical(1)=TRUE]} \cr If \code{TRUE}, the variable importance scores are
dividing the the maximum score, so that the score of the most important variable
is set to 1.}
}
\description{
Plot variable importance
}
