% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_plot_response.R
\name{plot_response}
\alias{plot_response}
\title{Plot responses from a bag of models}
\usage{
plot_response(
  x,
  dfvar,
  data,
  type = c("linear", "exponential")[2],
  zoi_shape = c("exp_decay", "gaussian_decay", "linear_decay", "threshold_decay")[1],
  which_cumulative = "cumulative",
  ci = TRUE,
  wQ_probs = c(0.025, 0.5, 0.975),
  baseline = c("median", "mean", "zero")[1],
  zoi = FALSE,
  zoi_vals = c(100, 250, 500, 1000, 2500, 5000, 10000),
  ggplot = T,
  plot_mean = TRUE,
  plot_median = TRUE,
  n_features = 1,
  normalize = c(FALSE, "mean", "median", "ci")[1],
  logx = F,
  ylim = NULL
)
}
\arguments{
\item{x}{\verb{[bag,list]} \cr A bag of models, resulting from a call to \code{\link[=bag_models]{bag_models()}}.}

\item{dfvar}{\verb{[data.frame]} \cr A data.frame with the values of the variables one wants to vary.
All other variables are set to their mean or median (this is set by the parameter \code{baseline}).
The column names of the dataframe might correspond exactly to the model covariates or to
parts of that (for instance, "roads_paved_" to refer to all ZOI variables related to paved roads).}

\item{data}{\verb{[data.frame]} \cr The original data used for model fitting. Used only for
taking the categories of the categorical variables.}

\item{type}{\verb{[character(1)="linear"]\{"linear", "exponential"\}}}

\item{ci}{Should variation or confidence intervals be plotted?}

\item{wQ_probs}{a three element with lower, mid, and higher weighted quantiles to be computed}
}
\description{
Plot responses from a bag of models
}
