% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_coefficients.R
\name{rescale_coefficients}
\alias{rescale_coefficients}
\title{Rescale standardized coefficients back to their original range after model fitting}
\usage{
rescale_coefficients(model, data)
}
\description{
Rescale standardized coefficients back to their original range after model fitting
}
