% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_coefficients.R
\name{rescale_coefficients}
\alias{rescale_coefficients}
\alias{rescale_coefficients.coxph}
\alias{rescale_coefficients.lm}
\alias{rescale_coefficients.glm}
\alias{rescale_coefficients.bag}
\title{Rescale standardized coefficients back to their original range after model fitting}
\usage{
rescale_coefficients(...)

\method{rescale_coefficients}{coxph}(model, data, ...)

\method{rescale_coefficients}{lm}(model, data, ...)

\method{rescale_coefficients}{glm}(model, data, ...)

\method{rescale_coefficients}{bag}(bag, data, tostd = TRUE, ...)
}
\arguments{
\item{model}{\cr Fitted model, i.e. the object created by a model fit funtion
such as "lm", "glm", or "coxph".}

\item{data}{\cr The original \code{data.frame} with the data used to fit the model.}

\item{bag}{\cr A bag of models, as result of the \code{\link[=bag_models]{bag_models()}} function.}

\item{standardize}{\verb{[logical(1)=TRUE]} If \code{TRUE} (Default), the coefficients are standardized.
If \code{FALSE}, the coefficients are standardized. Only numeric coefficients are
standardized.}
}
\value{
A vector of rescaled coefficients for the input model.
}
\description{
Predictor variables are often standardized to be included in statistical models
and allow comparison of the effect sizes for different predictors. This functions
scales the fitted models coefficients back to the original scale of the predictors,
to allow ecological interpretation.
}
\examples{
library(dplyr)

# standardize predictors
iris_std <- iris |>
  dplyr::mutate(across(2:4, ~ scale(.x)))
# fit model
m1 <- lm(Sepal.Length ~ Petal.Length + Species, data = iris_std)
summary(m1)

# rescale coefficients
(resc_cf <- rescale_coefficients(m1, iris))

# compare with model with no standardization of predictors
coef(lm(Sepal.Length ~ Petal.Length + Species, data = iris))

}
