% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sample_area.gpkg}
\alias{sample_area.gpkg}
\title{Sample area: a polygon vector data}
\format{
A geopackage file. Projected CRS: ETRS89 / UTM zone 33N.
}
\description{
Dataset containing the limits of an arbitrary study area in Southern Norway, used
for illustrative purposes.
}
\examples{
(f <- system.file("vector/sample_area.gpkg", package = "oneimpact"))
sf::st_read(f)
# or
v <- terra::vect(f)
plot(v)

}
\seealso{
Maps for the sample area: \cr
Cabins: \link{sample_area_cabins.gpkg}, \link{sample_area_cabins.tif},
\link{sample_area_cabins_count.tif} \cr
Roads: \link{sample_area_roads.gpkg}, \link{sample_area_roads.tif}
}
