% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sample_area_roads.gpkg}
\alias{sample_area_roads.gpkg}
\title{Road vector data for the sample area}
\format{
A geopackage file. Projected CRS: ETRS89 / UTM zone 33N. The vector
file presents the following columns:
\itemize{
\item{id:} {Line number, corresponding to the original dataset}
\item{name:} {Local name of the road}
\item{publ_priv:} {Whether the road is public or private}
\item{traffic_bin:} {Binary classification of the traffic on the road - high or low}
\item{name_area:} {Name of the reindeer management area where the road is located}
\item{traffic_bin:} {Value 1, to be used for rasterization purposes}
}
}
\source{
\url{https://kartkatalog.geonorge.no/metadata/elveg/ed1e6798-b3cf-48be-aee1-c0d3531da01a}
}
\description{
Dataset containing the location of public roads in Southern
Norway, within the sample area for the \code{oneimpact} package.
Retrieved from the Norwegian road dataset Elveg 1.0.
}
\examples{
(s <- system.file("vector/sample_area_roads.gpkg", package = "oneimpact"))
sf::st_read(s)
# or
terra::vect(s)

}
\seealso{
Maps for the sample area: \cr
Limits of sample area: \link{sample_area.gpkg} \cr
Cabins: \link{sample_area_cabins.gpkg}, \link{sample_area_cabins.tif},
\link{sample_area_cabins_count.tif} \cr
Roads: \link{sample_area_roads.tif}
}
