% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_mfilter.R
\name{save_mfilter}
\alias{save_mfilter}
\title{Save kernel/filter matrix to use in r.mfilter within GRASS GIS}
\usage{
save_mfilter(
  filt,
  zoi,
  method,
  save_format = c("GRASS", "raw")[1],
  save_folder = NULL,
  save_file = NULL,
  normalize = FALSE,
  parallel = TRUE,
  separator = " "
)
}
\arguments{
\item{filt}{\verb{[matrix]} \cr Matrix, such as one created by \code{\link[=create_filter]{create_filter()}} or \code{\link[terra:focalMat]{terra::focalMat()}}.}

\item{zoi}{\verb{[numeric(1)]} \cr Zone of Influence of the matrix, in meters.}

\item{method}{\verb{[character(1)]} Function for the kernel or filter matrix (see \code{method} parameter for \code{\link[=create_filter]{create_filter()}}).}

\item{save_format}{\verb{[character(1)=GRASS]\{"GRASS", "raw"\}} \cr Format in which the function should be saved. Currently, only GRASS GIS format
(\code{save_format = "GRASS"}, according to the required format for \code{r.mfilter} module, details
\href{https://grass.osgeo.org/grass78/manuals/r.mfilter.html}{here}) or raw (\code{save_format = "raw"}),
in which only the values of the matrix are printed.}

\item{save_folder}{\verb{[character(1)=NULL]} \cr Path to the folder where the matrix file should be written.
If \code{NULL}, the current directory is used.}

\item{save_file}{\verb{[character(1)=NULL]} \cr Name of the output file, generally a ".txt" file.
If \code{NULL}, a standard filename is created, using the the \code{method} and \code{zoi}. E.g. "filter_bartlett2000.txt".}

\item{normalize}{\verb{[logical(1)=FALSE]} \cr Whether the matrix should be normalized (sum of all cell is 1, if
\code{normalize = TRUE}) or kept as it is (default, \code{normalize = FALSE}).}

\item{parallel}{\verb{[logical(1)=TRUE]} \cr Whether the computation should be paralelized or not (details in
the documentation of the \href{(https://grass.osgeo.org/grass78/manuals/r.mfilter.html)}{\code{r.mfilter}} module).}

\item{separator}{\verb{[character(1)=" "]} \cr Separator between values of the matrix, within each line. Default is
a space.}
}
\value{
None. The funcion only saves the input matrix as an external file.
}
\description{
Save kernel/filter matrix to use in r.mfilter within GRASS GIS
}
\examples{
my_filter <- create_filter(r = 100, method = "bartlett", zoi = 1000, round = 4)
save_mfilter(my_filter, zoi = 1000, method = "bartlett", save_format = "GRASS")

}
