% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_points.R
\name{set_points}
\alias{set_points}
\title{Simulate points in a landscape}
\usage{
set_points(
  n_features = 1000,
  centers = 1,
  width = 0.05,
  res = 0.1,
  extent_x = c(0, 1),
  extent_y = c(0, 1),
  buffer_around = 0
)
}
\arguments{
\item{n_features}{\verb{[integer(1)=1000]} \cr Total number of features to spread in space.}

\item{centers}{\verb{[integer(1)=1]} \cr Number of centers around which the features will be placed.}

\item{width}{\verb{[numeric(1)=0.05]} \cr Radius of the "patches" of features, around the "patch" centers}

\item{res}{\verb{[numeric(1)=0.1]} \cr Resolution of the output raster.}

\item{extent_x, entent_y}{\verb{[numeric vector(2)=c(0,1)]} \cr Vector representing the minimum and
maximum extent in x and y within which the points should be placed, in the format c(min,max).}

\item{buffer_around}{\verb{[numeric(1)=0.1]} \cr Size of the buffer around the extent of the landscape,
to avoid edge effects when calculating densities using neighborhood analysis.}
}
\value{
A list with two elements: (1) \code{pts}, the coordinates (x,y) of the simulated points;
(2) \code{rast}, a binary raster containing the landscape, with 1 where there points and NA elsewhere.
}
\description{
This function simulate points patterns in space and rasterize them.
The idea is to mimic the spatial distribution of point-type infrastructure,
such as houses, cabins, or turbines, for instance.
The function returns a list with the position of the points and a binary raster
with 1 where there are points and NA elsewhere.
}
\details{
The function builds upon the function \code{\link[mobsim:sim_thomas_community]{mobsim::sim_thomas_community()}} from the \link{mobsim}
package. Originally the function is intended to simulate positions of multiple
species in the context of species abundance distribution studies, but it fits
well in case of a single species (or point patterns for a single type of feature).

TO IMPROVE: implement rasterization with terra package
}
\examples{
library(mobsim)
library(raster)

set.seed(1234)

# gradient distribution
ext <- 30000
wd <- ext/5
pts <- set_points(n_features = 1000, centers = 1,
                  width = wd, res = 100,
                  extent_x = c(0, ext), extent_y = c(0, ext))
plot(pts$pts)
plot(pts$rast)

# one focus of features, with buffer around
wd <- ext/20
pts <- set_points(n_features = 1000, centers = 1,
                  width = wd, res = 100,
                  extent_x = c(0, ext), extent_y = c(0, ext),
                  buffer_around = 10000)
plot(pts$pts)
plot(pts$rast)
}
