% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_points_sample.R
\name{set_points_sample}
\alias{set_points_sample}
\title{Simulate regular or random points in 2D}
\usage{
set_points_sample(
  n_features = 1000,
  type = c("regular", "random")[1],
  base_polygon = NULL,
  extent_x = c(0, 1),
  extent_y = extent_x
)
}
\arguments{
\item{n_features}{\verb{[integer(1)=1000]} \cr Total number of points to spread in space.}

\item{type}{\verb{[character(1)="regular"]\{"regular", "random"\}} Pattern for the creation
of points is space. Other methods are also accepted, check the \code{type} argument
for the \code{\link[sf:st_sample]{sf::st_sample()}} function.}

\item{base_polygon}{\verb{[RasterLayer,sfc_POLYGON]} \cr Polygon (from class \code{sf} or \code{sfc})
inside which the points will be created. If a \code{RasterLayer}, the \code{bbox} of the
raster is used as a polygon.}

\item{extent_x, entent_y}{\verb{[numeric vector(2)=c(0,1)]} \cr Vector representing the minimum and
maximum extent in x and y within which the points should be placed, in the format c(min,max).}
}
\value{
A \code{data.frame} with the (x,y) coordinates of the simulated points.
}
\description{
This function simulates the coordinates of points either regularly or
randomly  distributed in two dimensions. Other point patterns
can be also chosen, see the \code{type} argument for the \code{\link[sf:st_sample]{sf::st_sample()}}
function. The points are generated within an input polygon or the
bounding box of an input raster. If no spatial object is used as input,
The size of the landscape is defined by the \code{extent_x} and \code{extent_y} parameters.
It assumes the landscape is square or rectangularly shaped.
}
\examples{
library(sf)

pts <- set_points_sample(100)
plot(pts)
pts2 <- set_points_sample(100, type = "random")
plot(pts2)

library(raster)
x <- rast(system.file("external/test.grd", package="raster"))
pts3 <- set_points_sample(100, base_polygon = x)
plot(pts3)

}
