% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_resamples.R
\name{set_sampling_params}
\alias{set_sampling_params}
\title{Set sampling parameters}
\usage{
set_sampling_params(
  max_size_validate_blockH0 = 1000,
  max_number_blocksH1_train = 10,
  max_size_fit_blockH1 = 1000
)
}
\arguments{
\item{max_number_blocksH1_train}{Default = 10. Maximum number of blocks at level H1,
used for fitting the model. Why do we set a maximum here?
If all blocks should be used, use \code{max_number_blocksH1_train = Inf}.}

\item{max_validation_size_blockH0}{= 1000,}

\item{max_fit_size}{1000}
}
\value{
List of sampling parameters for the spatial stratified cross-validation,
with size of blocks of H0, H1, and H2.
}
\description{
This function sets the sampling parameters.
}
\examples{
set_sampling_params(1000, 10, 1000)
}
