% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance.R
\name{variable_importance}
\alias{variable_importance}
\title{Computes variable importance}
\usage{
variable_importance(
  x,
  data,
  type = c("drop", "permutation")[1],
  nb_permutations = 100,
  order = c("desc", "asc", FALSE)[1],
  plot = TRUE,
  remove_threshold = 0
)
}
\arguments{
\item{x}{\code{list} \cr Bag of models, result of \code{\link[=bag_models]{bag_models()}}. It contains
multiple information from the models, such as formula, weights, coefficients,
and metric used to evaluate the models.}
}
\description{
Computes variable importance
}
