% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoi_from_curve.R
\name{zoi_from_curve}
\alias{zoi_from_curve}
\alias{zoi_from_curve.data.frame}
\alias{zoi_from_curve.bag}
\title{Get estimates of zone of influence (ZOI) from response curves}
\usage{
zoi_from_curve(x, ...)

\method{zoi_from_curve}{data.frame}(
  x,
  percentage = 0.95,
  curve = c("median", "mean"),
  ci = TRUE,
  type = c("linear", "exp")[1],
  mean_col_name = "mean",
  median_col_name = "quantile:0.5",
  ci_col_name = c("quantile:0.025", "quantile:0.975")
)

\method{zoi_from_curve}{bag}(
  x,
  data,
  include = "all",
  percentage = 0.95,
  curve = c("median", "mean"),
  type = c("linear", "exp")[1],
  return_predictions = FALSE,
  return_format = c("list", "df")[2],
  ci = TRUE,
  wq_probs = c(0.025, 0.5, 0.975),
  n_features = 1,
  mean_col_name = "mean",
  median_col_name = "quantile:0.5",
  ci_col_name = c("quantile:0.025", "quantile:0.975"),
  radius_max = NULL,
  baseline = "zero",
  type_feature = "line",
  type_feature_recompute = TRUE,
  resolution = 200,
  radii = c(100, 250, 500, 1000, 2500, 5000, 10000),
  zoi_shape = c("circle", "Gauss", "rectangle", "exp_decay", "bartlett", "threshold",
    "mfilter")[1],
  ...
)
}
\arguments{
\item{x}{Either a \code{data.frame} containing response curve predictions for a single variable,
or a \code{bag} object containing an ensemble of models.}

\item{...}{Additional arguments passed to the appropriate method.}

\item{percentage}{\verb{[numeric(1)=0.95]} \cr Numeric between 0 and 1. Defines the
threshold for ZOI radius as a proportion of the maximum effect size.
Default is \code{0.95}.}

\item{curve}{\verb{[character(1)=c("mean", "median")]} \cr Character vector.
Which central tendency curves to use: \code{"median"}, \code{"mean"}, or both.}

\item{ci}{\verb{[logical(1)=TRUE]} \cr Logical. Whether to compute ZOI estimates for
the upper and lower limits of the confidence interval. Default is TRUE.}

\item{type}{\verb{[character(1)="linear"]\{"linear", "exp"\}} \cr Character. Defines whether
the calculation of ZOI should be based on the prediction of at linear
or response (exponential) scale: \code{"linear"} or \code{"exp"}, respectively.}

\item{mean_col_name}{\verb{[character="mean"]} \cr Name of the column containing
the mean response curve.}

\item{median_col_name}{\verb{[character="quantile:0.5"]} \cr Name of the column
containing the median response curve.}

\item{ci_col_name}{\verb{[character=c("quantile:0.255", "quantile:0.975")]} \cr Character
vector of length 2. Names of columns for lower and upper confidence intervals.}

\item{data}{\verb{[data.frame]} \cr The original dataset used for model fitting.}

\item{include}{\verb{[character="all"]} \cr Character. Either \code{"all"} or a
regex pattern to filter selected ZOI variables.}

\item{return_predictions}{\verb{[logical=FALSE]} \cr Logical. Whether to return
the prediction curves alongside ZOI metrics. If \code{TRUE}, the output is necessarily
a \code{list} with predictions and the ZOI parameters.}

\item{return_format}{\verb{[character="df"]\{"list", "df"\}} \cr
Format of the returned ZOI metrics. Either a list of data.frames (if \code{return_format = "list"}),
one for each variable, or a single \code{data.frame} (default, if \code{return_format = "df"}).}

\item{wq_probs}{\verb{[numeric,vector=c(0.025, 0.975)]} \cr Numeric vector of quantiles
used for prediction summaries.}

\item{n_features}{\verb{[numeric=1]} \cr Number of features used in ZOI prediction.
It can a single number (considered the same for all ZOI variables) or a vector
with the same number of elements as ZOI variables in the model.}

\item{radius_max}{\verb{[numeric=NULL]} \cr Numeric. Maximum distance/radius to use for
prediction curves. If \code{NULL} (default), the maximum value present in the bag's
predictor table is used.}

\item{baseline}{\verb{[character="zero"]} \cr Character. Baseline used in \code{predict()} (e.g., \code{"zero"}).}

\item{type_feature}{\verb{[character="point"]} \cr Character or vector. Type of spatial feature used in
\code{predict()}.}

\item{type_feature_recompute}{\verb{[logical=FALSE]} \cr Logical. Whether to recompute spatial
features within \code{predict()}, for linear features.}

\item{resolution}{\verb{[numeric=200]} \cr Integer. Resolution used in the recomuptation
of ZOIs for linear features.}

\item{radii}{\verb{[vector]} \cr Numeric vector. Radii used for ZOI modeling.}

\item{zoi_shape}{\verb{[character]} \cr Character. Shape of the ZOI used in the model
(e.g., \code{"circle"}, \code{"Gauss"}, \code{"exp_decay"}).}
}
\value{
A \code{data.frame} or a \code{list} containing ZOI metrics:
\itemize{
\item \code{max_effect_size}: Maximum effect size on the relative selection strength (y axis).
\item \code{zoi_radius}: Distance at which the effect drops below a threshold,
defined by the parameter \code{percentage}.
\item \code{effect_zoi_radius}: Relative selection strength (y axis) valye in which the ZOI
is reached.
\item \code{impact}: Area under the curve up to the ZOI radius,
combining the varying effect size with distance.
Each ZOI measure presents mean, median, CI lower, and CI upper.
}

If \code{x} is a bag object, the function returns wither a \code{list} or
\code{data.frame} of ZOI measures for each ZOI variable in the bag.
If \code{return_predictions = TRUE}, also returns the prediction curves.
}
\description{
This generic function computes ZOI metrics (maximum effect size,
ZOI radius and impact) for ZOI predictor variables based on response curves.
The ZOI radius is estimated as the distance/radius in which the
relative selection strength corresponds to on a given percentage of
the maximum effect size (e.g. 95\% ZOI radius). The impact accounts for
both the effect size and the ZOI radius and corresponds to the area under
(or over, if negative) the ZOI response curve.
The function supports two types of input: a data.frame of predictions
or a bag of models.
}
\seealso{
\code{\link[=predict]{predict()}}, \code{\link[=plot_response]{plot_response()}}, \code{\link[=weirdness]{weirdness()}}
##example examples/zoi_from_curve_example.R
}
