% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoi_functions.R
\name{zoi_functions}
\alias{zoi_functions}
\alias{dist_decay}
\alias{threshold_decay}
\alias{threshold_decay.numeric}
\alias{threshold_decay.SpatRaster}
\alias{step_decay}
\alias{bartlett_decay}
\alias{bartlett_decay.numeric}
\alias{bartlett_decay.SpatRaster}
\alias{tent_decay}
\alias{linear_decay}
\alias{gaussian_decay}
\alias{half_norm_decay}
\alias{exp_decay}
\title{Zone of Influence (ZoI) functions}
\usage{
dist_decay(
  x,
  radius = NULL,
  type = c("exp_decay", "gaussian_decay", "linear_decay", "threshold_decay")[1],
  zoi_limit = 0.05,
  origin = 0,
  oneside = TRUE,
  ...
)

threshold_decay(x, radius, constant_influence = 1, origin = 0, oneside = TRUE)

\method{threshold_decay}{numeric}(x, radius, constant_influence = 1, origin = 0, oneside = TRUE)

\method{threshold_decay}{SpatRaster}(x, radius, constant_influence = 1, origin = 0, oneside = TRUE)

step_decay(x, radius, constant_influence = 1, origin = 0, oneside = TRUE)

bartlett_decay(x, radius, intercept = 1, origin = 0, oneside = TRUE)

\method{bartlett_decay}{numeric}(x, radius, intercept = 1, origin = 0, oneside = TRUE)

\method{bartlett_decay}{SpatRaster}(x, radius, intercept = 1, origin = 0, oneside = TRUE)

tent_decay(x, radius, intercept = 1, origin = 0, oneside = TRUE)

linear_decay(x, radius, intercept = 1, origin = 0, oneside = TRUE)

gaussian_decay(
  x,
  radius = NULL,
  zoi_limit = 0.05,
  hnorm_decay_parms = c(1, 0.01),
  sigma = NULL,
  origin = 0,
  ...
)

half_norm_decay(
  x,
  radius = NULL,
  zoi_limit = 0.05,
  hnorm_decay_parms = c(1, 0.01),
  sigma = NULL,
  origin = 0,
  ...
)

exp_decay(
  x,
  radius = NULL,
  zoi_limit = 0.05,
  exp_decay_parms = c(1, 0.01),
  origin = 0,
  oneside = TRUE,
  half_life = NULL,
  zoi_hl_ratio = NULL
)
}
\arguments{
\item{x}{\verb{[numeric,SpatRaster,RasterLayer]} \cr Euclidean distance from an infrastructure, source
of disturbance, or feature/class of interest. It can be a single value, an array
of values, or a raster object. It must not necessarily be an Euclidean distance,
but preferably it should be a distance measured in meters, to ease interpretation
(e.g. geodesic distance).}

\item{radius}{\verb{[numeric(1)]} \cr Radius of the zone of influence (ZoI),
the distance at which the ZoI vanishes or goes below a given minimum limit value
\code{zoi_limit}. See details.}

\item{type}{\verb{[character(1)="Gauss"]\{"Gauss", "exp_decay", "bartlett", "linear", "tent", "threshold", "step"\}} \cr Type or shape of the decay distance.
\itemize{
\item If \code{"Gauss"} or \code{"half_norm"}, the ZoI follows a half-normal shape: \cr
\code{N_0 * exp(-lambda * (euclidean_distance^2))}. \code{N_0} and \code{lambda} are
parameters to be defined -- see details.
\item If \code{"exp_decay"}, the ZoI follows an exponential decay shape: \cr
\code{N_0 * exp(-lambda * euclidean_distance)}. \code{N_0} and \code{lambda} are
parameters to be defined -- see details.
\item If \code{"bartlett"}, \code{"linear_decay"}, or \code{"tent_decay"}, the ZoI follows a
linear decay shape within the ZoI radius (parameter \code{radius}).
\item If \code{"threshold"} or \code{"step"}, a constant influence is consider within the
zone of influence radius (parameter \code{radius}). All pixels closer than
\code{radius} to infrastructure are considered as "under the influence" of
the nearest feature, with a constant influence value defined by the
\code{constant_influence} parameter, and all other pixels are assumed to have
zero influence.
}}

\item{zoi_limit}{\verb{[numeric(1)=0.05]} \cr For non-vanishing functions
(e.g. \code{exp_decay}, \code{gaussian_decay}), this value is used to set the relationship
between the ZoI radius and the decay functions:
\code{radius} is defined as the minimum distance \code{x} at which the ZoI assumes values
below \code{zoi_limit}. The default is 0.05. This parameter is used only
if \code{radius} is not \code{NULL}.}

\item{origin}{\verb{[numeric(1)=0]} \cr In which position (in 1 dimension) is located
the infrastructure or source of disturbance? Default is zero. For raster objects,
this parameter should be ignored.}

\item{oneside}{\verb{[logical(1)=TRUE]} \cr If \code{FALSE}, negative distance values
are considered symmetrically and their transformation is always positive.
This parameter is only meaningful if \code{x} is a vector of values, not a
raster object.}

\item{constant_influence}{\verb{[numeric(1)=1]} \cr Constant value of the
\code{threshold} (or \code{step}) function within the Zone of Influence. Default is 1.}

\item{intercept}{\verb{[numeric(1)=1]} For the Bartlett (linear or tent decay) function,
\code{intercept} is the maximum value of the function (at \code{x = 0}).}

\item{hnorm_decay_parms}{\verb{[numeric(2)=c(1,0.01)]} \cr For the \code{gaussian_decay} function,
these are the guassian decay parameters c(N, lambda), where N is the maximum
value of the function (at \code{x = 0})
and lambda is the decay parameter of the Gaussian function.
The value for \code{lambda} is only considered to draw the Gaussian decay function
if both \code{radius = NULL} and \code{sigma = NULL}.}

\item{sigma}{\verb{[numeric(1)=NULL]} \cr Standard deviation of the Gaussian
function. It related to the Gaussian decay rate \eqn{\lambda} as
\code{lambda = 1/(2*sigma^2)}. Only considered to compute the ZoI
for the \code{gaussian_decay} function when the ZoI radius parameter is null
(\code{radius = NULL}).}

\item{exp_decay_parms}{\verb{[numeric(2)=c(1,0.01)]} \cr For the \code{exp_decay} function,
these are the exponential decay parameters \code{c(N_0, lambda)}, where \code{N_0} is the maximum
value of the function (at \code{x = 0})
and lambda is the decay parameter of the exponential function.
The value for \code{lambda} is only considered to draw the exponential decay function
if both \code{radius = NULL} and \code{half_life = NULL}.}

\item{half_life}{\verb{[numeric(1)=NULL]} \cr Half life of the exponential decay
function, in meters (or map units, for rasters). By definition, the half life is
the distance where the exponential decay function reaches 0.5 of its
maximum value. For the \code{exp_decay} function,
if the ZoI radius parameter is null (\code{radius = NULL}), the value of the
exponential half life (\code{half_life = log(2)/lambda}) can used to parameterize the
exponential decay function.}

\item{zoi_hl_ratio}{\verb{[numeric(1)=NULL]} \cr For the \code{exp_decay} function,
if both the ZoI radius \code{radius} and \code{zoi_hl_ratio} are given and
\code{half_life} is \code{NULL}, this value is used
to set the ZoI radius (and \code{zoi_limit} is ignored).
\code{zoi_hl_ratio} is the ratio between the
ZoI radius value and the half life of the exponential function.
For instance, if \code{radius = 1200} and \code{zoi_hl_ratio = 6}, this means
\code{half_life} is 200. As a consequence, the exponential decay ZoI function
decreases to 0.5 at distance 200, and the ZoI radius = 1200
is defined as the distance
at which the ZoI decreases to 0.5**6 = 0.015625.}
}
\value{
The ZoI values for a given array of distance values if \code{x} is numeric,
or a raster object delimiting the ZoI if \code{x} corresponds to the distance from
infrastructure or disturbance sources in 2 dimensions space.
}
\description{
Computes the decay functions that represent the Zone of
Influence (ZoI). The functions might assume different shapes
(parameter \code{type}) to represent multiple ways
the ZoI of an infrastructure or disturbance affects a
given process in space, and the ZoI radius (parameter \code{radius})
controls how far this effect reaches. The rate
of decay of the different ZoI functions is parameterized based on
the ZoI radius -- e.g the slope of \code{\link[=linear_decay]{linear_decay()}} is defined
so that the function decreases to zero at the ZoI radius.
These functions can be used to transform arrays of (Euclidean)
distance values (in one dimension) or rasters of (Euclidean) distance
(in two dimensions) into ZoI values. The distances might represent
the distance to human infrastructure, sources of disturbance, or
more broadly any type of land use class or spatial variable.
}
\details{
A generic function \code{\link[=dist_decay]{dist_decay()}} can be used to compute
ZoI values according to functions with different shapes (parameter \code{type})
and radii (parameter \code{radius}). Alternatively, there are specific functions
implemented for each ZoI shape.

For the threshold function (\code{\link[=threshold_decay]{threshold_decay()}}) and the linear decay
function (\code{\link[=linear_decay]{linear_decay()}}), the ZoI radius (parameter \code{radius}) is the
distance \code{x} where the ZoI function value decreases to zero.
For the linear decay, this is done by setting
the slope of the linear function as \code{-intercept/radius}, where \code{intercept}
is the intercept of the linear function (here, the maximum value at \code{x = 0}).

For non-vanishing functions that approach zero asymptotically
(\code{\link[=exp_decay]{exp_decay()}}, \code{\link[=gaussian_decay]{gaussian_decay()}}), a certain limit value must be given to define
the ZoI radius -- so that the ZoI radius is defined as the distance \code{x} where the
ZoI function goes below this limit value. For these functions,
different parameters are available
for setting the relationship between the ZoI function value and the ZoI radius.

Some functions have multiple possible names, for the sake of flexibility:
\itemize{
\item \code{\link[=linear_decay]{linear_decay()}}, \code{\link[=bartlett_decay]{bartlett_decay()}}, and
\code{\link[=tent_decay]{tent_decay()}} are the same function;
\item \code{\link[=threshold_decay]{threshold_decay()}} and \code{\link[=step_decay]{step_decay()}} are the same function;
\item \code{\link[=gaussian_decay]{gaussian_decay()}} and \code{\link[=half_norm_decay]{half_norm_decay()}} are the same function.
}

Alternatively, \code{\link[=dist_decay]{dist_decay()}} can call all of them, given a
ZoI shape is specified through the parameter \code{type}.

Other functions might be implemented in the future.
}
\section{Definitions}{
Here are some formal definitions for the ZoI functions \eqn{\phi(d_i, r)},
where \eqn{d_i} is the distance to the feature \eqn{i} of an infrastructure or
source of disturbance and \eqn{r} is the radius of the zone of influence:
\itemize{
\item \code{threshold_decay()}: the threshold or step decay function \eqn{\phi_{threshold}} is
positive and constant within the ZoI radius \eqn{r}, and null for \eqn{x \ge r}:
\deqn{
\phi_{threshold}(d_i, r_k) = c \text{ if } d_i < r, 0 \text{ otherwise}
}
where \eqn{c} is a constant value (by default \code{c = 1}).
\item \code{linear_decay()}: the linear (or tent/Bartlett) decay function \eqn{\phi_{linear}}
decreases
linearly from a maximum value \eqn{b} (the intercept, by default \code{b = 1}) to
zero when \eqn{x \ge r}:
\deqn{\phi_{linear}(d_i, r) = b - b/r \text{ if } x < r, \text{ 0 otherwise}}
\item \code{exp_decay()}: the exponential decay function \eqn{\phi_{exp}} decreases
exponentially from a maximum value \eqn{N} (by default \code{N = 1}) with a rate
\eqn{\lambda}, which is defined by \eqn{r} and a ZoI limit value
\eqn{\phi_{lim}}, a small ZoI value below which the effect is considered negligible:
\deqn{\phi_{exp}(d_i, r, \phi_{lim}) = N exp(-\lambda d_i)}
with
\deqn{\lambda = ln(1/\phi_{lim}) / r}
In this context, the ZoI radius \eqn{r} is the distance beyond which
\eqn{\phi_{exp} < \phi_{lim}}.
\item \code{gaussian_decay()}: the Gaussian decay function \eqn{\phi_{Gauss}}
follows a Gaussian (half-normal) decay with maximum \eqn{N} (by default \code{N = 1})
and a decay rate \eqn{\lambda}
defined by \eqn{r} and a ZoI limit value
\eqn{\phi_{lim}}, a small ZoI value below which the effect is considered negligible:
\deqn{\phi_{Gauss}(d_i, r, \phi_{lim}) = N exp(-\lambda d_i^2)}
with
\deqn{\lambda = ln(1/\phi_{lim}) / (r^2)}
In this context, the ZoI radius \eqn{r} is the distance beyond which
\eqn{\phi_{exp} < \phi_{lim}}. Note that \eqn{\lambda} is defined differently
for the \code{gaussian_decay} and the \code{exp_decay} functions.
}
}

\section{Parameterization}{
Some of the shapes of the ZoI (parameter \code{type} in \code{dist_decay()}) might be
parameterized in multiple ways. Here is a brief description of each possibility:

\itemize{

\item For the \code{"Gauss"} or \code{"half_norm"} shapes, the ZoI follows a half-normal shape: \cr
\code{N_0 * exp(-lambda * (euclidean_distance^2))}. \code{N_0} and \code{lambda} are
parameters to be defined. \code{N_0} is always defined as the first element of the
parameter \code{hnorm_decay_parms}. There are three ways of specifying \code{lambda}:
\itemize{
\item If the \code{radius = NULL} (default), \code{lambda} is defined as the second
element of the parameter \code{hnorm_decay_parms}. In all other cases (below)
this value is ignored.
\item If the parameter \code{radius} is given, the rate of decay is given by
\code{lambda = log(1/zoi_limit) / (radius**2)}. In other words, \code{lambda} is defined
so that the function decreases to \code{zoi_limit} when \code{x = radius}.
\item If the \code{radius = NULL} and \code{sigma} is given, \code{lambda} is defined as
\code{lambda = 1/(2*sigma**2)}.
}

\item For the \code{"exp_decay"} shape, the ZoI follows an exponential decay shape: \cr
\code{N_0 * exp(-lambda * euclidean_distance)}. \code{N_0} and \code{lambda} are
parameters to be defined. \code{N_0} is always defined as the first element of the
parameter \code{exp_decay_parms}.There are four ways of specifying \code{lambda}:
\itemize{
\item If the \code{radius = NULL} (default), \code{lambda} is defined as the second
element of the parameter \code{exp_decay_parms}. In all other cases (below)
this value is ignored.
\item If the parameter \code{radius} is given, the rate of decay is given by
\code{lambda = log(1/zoi_limit) / radius}. In other words, \code{lambda} is defined
so that the function decreases to \code{zoi_limit} when \code{x = radius}.
\item If the \code{radius = NULL} and \code{half_life} is given, \code{lambda} is defined
based on the half life of the exponential function -- the distance at which
the function decreases to 1/2. If \code{zoi_hl_ratio = NULL}, \code{lambda} is defined
as \code{lambda = log(2)/half_life}.
\item The last possibility is to specify \code{zoi_hl_ratio}, the ratio between
the ZoI radius and the half life of the exponential function. For instance,
if \code{zoi_hl_ratio = 4}, this means the ZoI radius is defined as \code{4*half_life}.
If \code{zoi_hl_ratio} is given, the exponential \code{half_life} is defined based on
this parameter and \code{lambda} is defined accordingly, based on the relationship
above. In this case, \code{radius} is ignored, even if specified.
}

\item For the \code{"bartlett"}, \code{"linear_decay"}, or \code{"tent_decay"} shapes, the ZoI follows a
linear decay shape (\code{y = a*x + b}) within the ZoI radius (parameter \code{radius}).
The intercept of the linear function (\code{b}) is given by the parameter \code{intercept}
and the slope (\code{a}) is given by \code{-intercept/radius}.

\item For the \code{"threshold"} or \code{"step"} shapes, a constant influence is consider
within the
zone of influence radius (parameter \code{radius}). All pixels closer than
\code{radius} to infrastructure are considered as "under the influence" of
the nearest feature, with a constant influence value defined by the
\code{constant_influence} parameter, and all other pixels are assumed to have
zero influence.
}
}

\examples{
# test the zone of influence functions
# here we use ggplot() to illustrate the functions, to make the figures more
# widely reproducible
# to ease the plots, use the function plot_zoi1d()
library(ggplot2)

# exponential decay
exp_decay(10, radius = 30)

f1 <- ggplot(data.frame(x = c(0, 30)), aes(x = x)) +
  stat_function(fun = exp_decay, args = list(radius = 20)) +
  labs(x = "Distance", y = "Zone of Influence") +
  theme_bw()
f1

# exponential decay - two sides
f1_2 <- ggplot(data.frame(x = c(-30, 30)), aes(x = x)) +
  stat_function(fun = exp_decay,
                args = list(radius = 20, oneside = FALSE)) +
  labs(x = "Distance", y = "Zone of Influence") +
  theme_bw()
f1_2

# threshold
threshold_decay(5, radius = 10)
threshold_decay(10, radius = 10)

f2 <- ggplot(data.frame(x = c(0, 30)), aes(x = x)) +
  stat_function(fun = threshold_decay,
                args = list(radius = 20), linetype = 2) +
  labs(x = "Distance", y = "Zone of Influence") +
  theme_bw()
f2

# threshold - two sides
f2_2 <- ggplot(data.frame(x = c(-30, 50)), aes(x = x)) +
  stat_function(fun = threshold_decay,
                args = list(radius = 20, oneside = FALSE), linetype = 2) +
  labs(x = "Distance", y = "Zone of Influence") +
  theme_bw()
f2_2

# linear, tent, or bartlett decay
bartlett_decay(5, radius = 10)
bartlett_decay(8, radius = 10)

f3 <- ggplot(data.frame(x = c(0, 30)), aes(x = x)) +
  stat_function(fun = bartlett_decay, args = list(radius = 20), linetype = 3) +
  labs(x = "Distance", y = "Zone of Influence") +
  theme_bw()
f3

# linear, two sides
f3_3 <- ggplot(data.frame(x = c(-30, 40)), aes(x = x)) +
  stat_function(fun = bartlett_decay,
                args = list(radius = 20, origin = 10, oneside = FALSE), linetype = 3) +
  labs(x = "Distance", y = "Zone of Influence") +
  theme_bw()
f3_3

# guassian or half normal
gaussian_decay(5, sigma = 6)

f4 <- ggplot(data.frame(x = c(0, 30)), aes(x = x)) +
  stat_function(fun = gaussian_decay,
                args = list(radius = 20, zoi_limit = 0.05), linetype = 4) +
  labs(x = "Distance", y = "Zone of Influence") +
  geom_vline(xintercept = 20, linetype = 2, color = "darkgrey") +
  geom_hline(yintercept = 0.05, linetype = 2, color = "darkgrey") +
  theme_bw()
f4

# half normal - two sides
gaussian_decay(5, sigma = 6)

f4_2 <- ggplot(data.frame(x = c(-30, 30)), aes(x = x)) +
  stat_function(fun = gaussian_decay,
                args = list(radius = 20, zoi_limit = 0.05), linetype = 4) +
  labs(x = "Distance", y = "Zone of Influence") +
  geom_vline(xintercept = c(-20, 20), linetype = 2, color = "darkgrey") +
  geom_hline(yintercept = 0.05, linetype = 2, color = "darkgrey") +
  theme_bw()
f4_2

# plot several ZoI with the same radius
f1 +
  stat_function(fun = threshold_decay, args = list(radius = 20), linetype = 2) +
  stat_function(fun = bartlett_decay, args = list(radius = 20), linetype = 3) +
  stat_function(fun = gaussian_decay, args = list(radius = 20, zoi_limit = 0.05), linetype = 4) +
  labs(x = "Distance", y = "Zone of Influence") +
  theme_bw()

# generic dist_decay function
dist_decay(500, radius = 1000, type = "exp_decay")
dist_decay(500, radius = 1000, type = "gaussian_decay")
dist_decay(500, radius = 1000, type = "linear_decay")
dist_decay(500, radius = 1000, type = "step_decay")

#---
# applying dist_decay functions for rasters
library(terra)

# calculate Euclidean distance
f <- system.file("raster/cabins.tif", package = "oneimpact")
cabins <- terra::rast(f)
cabins_dist <- calc_zoi_nearest(cabins, type = "euclidean")

# transform Euclidean in distance decay
plot(dist_decay(cabins_dist, radius = 1000, type = "exp_decay"))
plot(dist_decay(cabins_dist, radius = 1000, type = "tent_decay"))
}
